Bootsrap-enhanced Lasso

This function implements model-consistent Lasso estimation through the bootstrap. It supports parallel processing by way of the future package, allowing the user to flexibly specify many parallelization methods. This method was developed as a variable-selection algorithm, but this package also supports making ensemble predictions on new data using the bagged Lasso models.

Usage

bolasso(
  form,
  data,
  n.boot = 100,
  progress = TRUE,
  implement = "glmnet",
  x = NULL,
  y = NULL,
  ...
)

Arguments

form

An optional object of class formula (or one that can be coerced to that class): a symbolic description of the model to be fitted. Can be omitted when x and y are non-missing.

data

An optional object of class data.frame that contains the modeling variables referenced in form. Can be omitted when x and y are non-missing.

n.boot

An integer specifying the number of bootstrap replicates.

progress

A boolean indicating whether to display progress across bootstrap folds.

implement

A character; either 'glmnet' or 'gamlr', specifying which Lasso implementation to utilize. For specific modeling details, see glmnet::cv.glmnet or gamlr::cv.gamlr.

x

An optional predictor matrix in lieu of form and data.

y

An optional response vector in lieu of form and data.

...

Additional parameters to pass to either glmnet::cv.glmnet or gamlr::cv.gamlr.

Value

An object of class bolasso. This object is a list of length n.boot of cv.glmnet or cv.gamlr objects.

References

Bach FR (2008). “Bolasso: model consistent Lasso estimation through the bootstrap.” CoRR, abs/0804.1302. 0804.1302, https://arxiv.org/abs/0804.1302.

See also

Examples

mtcars[, c(2, 10:11)] <- lapply(mtcars[, c(2, 10:11)], as.factor)
idx <- sample(nrow(mtcars), 22)
mtcars_train <- mtcars[idx, ]
mtcars_test <- mtcars[-idx, ]

## Formula Interface

# Train model
set.seed(123)
bolasso_form <- bolasso(
  form = mpg ~ .,
  data = mtcars_train,
  n.boot = 20,
  nfolds = 5,
  implement = "glmnet"
)
#> Loaded glmnet 4.1-3

# Extract selected variables
selected_vars(bolasso_form, threshold = 0.9, select = "lambda.min")
#> # A tibble: 2 × 2
#>   variable  mean_coef
#>   <chr>         <dbl>
#> 1 Intercept     11.7 
#> 2 wt            -3.18

# Bagged ensemble prediction on test data
predict(bolasso_form,
        new.data = mtcars_test,
        select = "lambda.min")
#>                      boot1    boot2    boot3    boot4    boot5    boot6
#> Mazda RX4         23.26766 21.46072 20.41443 20.91785 23.62603 24.83718
#> Hornet Sportabout 19.31635 19.10984 21.04466 18.17266 19.41161 18.61427
#> Merc 240D         22.73438 26.81913 18.09434 25.02151 24.68292 23.92291
#> Merc 230          22.36888 25.93944 30.58464 24.80125 26.09336 22.86140
#> Merc 280          18.84694 19.18335 15.43930 18.68182 21.98044 19.08743
#> Merc 450SL        17.74597 17.44931 14.00087 17.37792 18.62414 17.15167
#> Dodge Challenger  19.33894 19.68858 19.90203 18.21171 18.60368 19.32233
#> AMC Javelin       19.77390 19.89173 22.72100 18.43030 19.79012 19.48671
#> Ford Pantera L    19.16020 15.43573 26.26245 17.99554 20.37869 18.45885
#> Ferrari Dino      21.19071 20.64472 21.50716 19.89565 21.60812 22.89607
#>                      boot7    boot8    boot9   boot10   boot11   boot12
#> Mazda RX4         22.04607 21.21836 18.50699 24.29169 20.82601 24.68322
#> Hornet Sportabout 19.11169 15.73753 20.14150 18.05702 20.85490 15.58704
#> Merc 240D         18.71824 22.00404 27.28882 22.37491 24.22367 21.06155
#> Merc 230          19.05419 21.93072 31.10573 21.47377 31.98479 25.25918
#> Merc 280          15.23155 18.81447 17.88809 19.66603 16.81430 18.85253
#> Merc 450SL        20.45477 16.61329 18.45985 16.97778 19.50161 15.39972
#> Dodge Challenger  18.88436 16.40832 19.66339 18.58146 19.47072 14.36036
#> AMC Javelin       19.14719 16.69805 20.51413 18.85211 21.57626 15.94671
#> Ford Pantera L    23.88388 16.96981 14.29360 17.74070 14.01015 21.62895
#> Ferrari Dino      26.91891 20.88360 20.37968 21.79988 18.11646 22.52250
#>                     boot13    boot14   boot15   boot16   boot17   boot18
#> Mazda RX4         23.63524 19.989828 21.72591 21.37257 18.13922 23.61170
#> Hornet Sportabout 18.18185 12.861407 18.92991 18.43125 22.88893 18.07862
#> Merc 240D         22.84111 20.968823 25.77361 21.74555 28.62094 25.77640
#> Merc 230          22.18168 31.892238 31.45393 21.84396 45.04634 26.67576
#> Merc 280          18.82139 22.700631 19.35080 19.47076 14.75017 21.89670
#> Merc 450SL        16.91529 17.028361 15.97126 16.07008 30.95173 17.44393
#> Dodge Challenger  18.57632  9.050242 17.88010 18.19823 25.11161 17.98826
#> AMC Javelin       18.75609 13.742110 19.44956 18.47355 29.49370 18.37707
#> Ford Pantera L    18.99981 21.081433 17.42431 17.49151 10.56418 15.14191
#> Ferrari Dino      21.85246 17.173132 19.17920 20.88984 27.52876 19.39544
#>                     boot19   boot20
#> Mazda RX4         21.76410 22.29269
#> Hornet Sportabout 18.71738 18.65455
#> Merc 240D         23.71424 22.02639
#> Merc 230          22.78992 21.53241
#> Merc 280          18.43975 18.46901
#> Merc 450SL        17.01712 17.29413
#> Dodge Challenger  19.36443 18.81298
#> AMC Javelin       19.69618 19.18208
#> Ford Pantera L    17.90502 16.97246
#> Ferrari Dino      21.99839 21.74769

## Alternal Matrix Interface

# Train model
set.seed(123)
bolasso_mat <- bolasso(
  x = model.matrix(mpg ~ . - 1, mtcars_train),
  y = mtcars_train[, 1],
  data = mtcars_train,
  n.boot = 20,
  nfolds = 5,
  implement = "glmnet"
)

# Extract selected variables
selected_vars(bolasso_mat, threshold = 0.9, select = "lambda.min")
#> # A tibble: 2 × 2
#>   variable  mean_coef
#>   <chr>         <dbl>
#> 1 Intercept     11.7 
#> 2 wt            -3.18

# Bagged ensemble prediction on test data
predict(bolasso_mat,
        new.data = model.matrix(mpg ~ . - 1, mtcars_test),
        select = "lambda.min")
#>                      boot1    boot2    boot3    boot4    boot5    boot6
#> Mazda RX4         23.26766 21.46072 20.41443 20.91785 23.62603 24.83718
#> Hornet Sportabout 19.31635 19.10984 21.04466 18.17266 19.41161 18.61427
#> Merc 240D         22.73438 26.81913 18.09434 25.02151 24.68292 23.92291
#> Merc 230          22.36888 25.93944 30.58464 24.80125 26.09336 22.86140
#> Merc 280          18.84694 19.18335 15.43930 18.68182 21.98044 19.08743
#> Merc 450SL        17.74597 17.44931 14.00087 17.37792 18.62414 17.15167
#> Dodge Challenger  19.33894 19.68858 19.90203 18.21171 18.60368 19.32233
#> AMC Javelin       19.77390 19.89173 22.72100 18.43030 19.79012 19.48671
#> Ford Pantera L    19.16020 15.43573 26.26245 17.99554 20.37869 18.45885
#> Ferrari Dino      21.19071 20.64472 21.50716 19.89565 21.60812 22.89607
#>                      boot7    boot8    boot9   boot10   boot11   boot12
#> Mazda RX4         22.04607 21.21836 18.50699 24.29169 20.82601 24.68322
#> Hornet Sportabout 19.11169 15.73753 20.14150 18.05702 20.85490 15.58704
#> Merc 240D         18.71824 22.00404 27.28882 22.37491 24.22367 21.06155
#> Merc 230          19.05419 21.93072 31.10573 21.47377 31.98479 25.25918
#> Merc 280          15.23155 18.81447 17.88809 19.66603 16.81430 18.85253
#> Merc 450SL        20.45477 16.61329 18.45985 16.97778 19.50161 15.39972
#> Dodge Challenger  18.88436 16.40832 19.66339 18.58146 19.47072 14.36036
#> AMC Javelin       19.14719 16.69805 20.51413 18.85211 21.57626 15.94671
#> Ford Pantera L    23.88388 16.96981 14.29360 17.74070 14.01015 21.62895
#> Ferrari Dino      26.91891 20.88360 20.37968 21.79988 18.11646 22.52250
#>                     boot13    boot14   boot15   boot16   boot17   boot18
#> Mazda RX4         23.63524 19.989828 21.72591 21.37257 18.13922 23.61170
#> Hornet Sportabout 18.18185 12.861407 18.92991 18.43125 22.88893 18.07862
#> Merc 240D         22.84111 20.968823 25.77361 21.74555 28.62094 25.77640
#> Merc 230          22.18168 31.892238 31.45393 21.84396 45.04634 26.67576
#> Merc 280          18.82139 22.700631 19.35080 19.47076 14.75017 21.89670
#> Merc 450SL        16.91529 17.028361 15.97126 16.07008 30.95173 17.44393
#> Dodge Challenger  18.57632  9.050242 17.88010 18.19823 25.11161 17.98826
#> AMC Javelin       18.75609 13.742110 19.44956 18.47355 29.49370 18.37707
#> Ford Pantera L    18.99981 21.081433 17.42431 17.49151 10.56418 15.14191
#> Ferrari Dino      21.85246 17.173132 19.17920 20.88984 27.52876 19.39544
#>                     boot19   boot20
#> Mazda RX4         21.76410 22.29269
#> Hornet Sportabout 18.71738 18.65455
#> Merc 240D         23.71424 22.02639
#> Merc 230          22.78992 21.53241
#> Merc 280          18.43975 18.46901
#> Merc 450SL        17.01712 17.29413
#> Dodge Challenger  19.36443 18.81298
#> AMC Javelin       19.69618 19.18208
#> Ford Pantera L    17.90502 16.97246
#> Ferrari Dino      21.99839 21.74769