Bootsrap-enhanced Lasso

This function implements model-consistent Lasso estimation through the bootstrap. It supports parallel processing by way of the future package, allowing the user to flexibly specify many parallelization methods. This method was developed as a variable-selection algorithm, but this package also supports making ensemble predictions on new data using the bagged Lasso models.

bolasso(
  form,
  data,
  n.boot = 100,
  progress = TRUE,
  implement = "glmnet",
  x = NULL,
  y = NULL,
  ...
)

Arguments

form

An optional object of class formula (or one that can be coerced to that class): a symbolic description of the model to be fitted. Can be omitted when x and y are non-missing.

data

An optional object of class data.frame that contains the modeling variables referenced in form. Can be omitted when x and y are non-missing.

n.boot

An integer specifying the number of bootstrap replicates.

progress

A boolean indicating whether to display progress across bootstrap folds.

implement

A character; either 'glmnet' or 'gamlr', specifying which Lasso implementation to utilize. For specific modeling details, see glmnet::cv.glmnet or gamlr::cv.gamlr.

x

An optional predictor matrix in lieu of form and data.

y

An optional response vector in lieu of form and data.

...

Additional parameters to pass to either glmnet::cv.glmnet or gamlr::cv.gamlr.

Value

An object of class bolasso. This object is a list of length n.boot of cv.glmnet or cv.gamlr objects.

References

Bach FR (2008). “Bolasso: model consistent Lasso estimation through the bootstrap.” CoRR, abs/0804.1302. 0804.1302, http://arxiv.org/abs/0804.1302.

See also

Examples

mtcars[, c(2, 10:11)] <- lapply(mtcars[, c(2, 10:11)], as.factor)
idx <- sample(nrow(mtcars), 22)
mtcars_train <- mtcars[idx, ]
mtcars_test <- mtcars[-idx, ]

## Formula Interface

# Train model
set.seed(123)
bolasso_form <- bolasso(
  form = mpg ~ .,
  data = mtcars_train,
  n.boot = 20,
  nfolds = 5,
  implement = "glmnet"
)
#> Loaded glmnet 4.1-3

# Extract selected variables
selected_vars(bolasso_form, threshold = 0.9, select = "lambda.min")
#> # A tibble: 2 × 2
#>   variable  mean_coef
#>   <chr>         <dbl>
#> 1 Intercept     21.5 
#> 2 wt            -3.04

# Bagged ensemble prediction on test data
predict(bolasso_form,
        new.data = mtcars_test,
        select = "lambda.min")
#>                       boot1     boot2    boot3    boot4    boot5    boot6
#> Valiant            20.63691 20.909550 19.82558 20.56410 19.73322 21.28793
#> Duster 360         15.76638 16.372454 14.02662 14.92655 15.52031 14.71217
#> Merc 230           25.12386 25.045148 22.92836 28.99311 21.73280 29.00448
#> Cadillac Fleetwood 11.87418  8.877457 12.36704 12.17847 12.54445 11.29728
#> Chrysler Imperial  11.09671  7.844000 11.18548 11.91890 11.61370 10.64259
#> Toyota Corolla     27.48744 29.725699 26.67354 31.37841 25.77683 31.20256
#> Toyota Corona      25.39340 26.295158 24.17597 30.16829 23.34638 27.62056
#> Fiat X1-9          26.47960 28.270351 26.43986 31.20191 25.50514 29.56025
#> Lotus Europa       26.33612 28.843855 25.38238 31.75100 25.21916 28.17203
#> Ferrari Dino       19.44345 20.549375 18.36977 21.48808 19.45716 19.70001
#>                       boot7    boot8    boot9   boot10   boot11    boot12
#> Valiant            20.88968 16.45979 21.81642 20.31133 18.31833 19.069060
#> Duster 360         16.12372 11.79266 14.05474 16.72145 16.61016 12.817112
#> Merc 230           25.87340 24.43973 31.50252 21.73801 21.90245 25.168089
#> Cadillac Fleetwood 11.68343 11.25757 11.60613 11.56359 10.84305  7.246436
#> Chrysler Imperial  10.71432 11.95087 11.02088 10.65026 10.32888  8.536854
#> Toyota Corolla     29.79012 27.14822 31.23569 27.17332 27.07122 24.570639
#> Toyota Corona      27.18482 20.52158 27.56142 24.16737 24.30330 26.303559
#> Fiat X1-9          28.79208 26.67265 28.51022 26.78921 26.65702 23.612131
#> Lotus Europa       28.00866 23.59150 25.15139 27.24218 27.61397 24.039371
#> Ferrari Dino       19.62817 20.66017 17.55627 21.20814 20.35387 20.173759
#>                      boot13   boot14    boot15    boot16   boot17   boot18
#> Valiant            20.96329 20.61230 22.217240 19.956577 18.65454 20.70182
#> Duster 360         11.75218 14.71489 14.937807 14.862105 16.51094 15.13084
#> Merc 230           29.39846 21.07515 25.881144 23.071813 21.66453 22.66934
#> Cadillac Fleetwood 14.80662 11.04649  7.117202 10.712186 11.63846 12.33739
#> Chrysler Imperial  14.51778 10.83905  5.906395  9.706662 11.01416 11.14948
#> Toyota Corolla     29.30298 23.94654 32.161672 26.195683 26.26276 27.17213
#> Toyota Corona      23.81786 22.57089 28.447409 23.310494 23.83429 24.34895
#> Fiat X1-9          27.27109 23.72819 30.423829 25.853065 25.92948 26.87812
#> Lotus Europa       22.72957 25.35814 29.083474 27.708635 26.68066 26.18377
#> Ferrari Dino       18.76525 24.25682 21.094349 20.146225 19.97039 19.82523
#>                      boot19    boot20
#> Valiant            19.74255 21.273184
#> Duster 360         16.04787 15.035304
#> Merc 230           23.13767 22.612717
#> Cadillac Fleetwood 12.57743 10.845459
#> Chrysler Imperial  11.85906  9.855632
#> Toyota Corolla     27.02833 27.582499
#> Toyota Corona      24.82676 24.713794
#> Fiat X1-9          26.75701 27.238644
#> Lotus Europa       26.92103 27.270005
#> Ferrari Dino       20.01838 21.511907

## Alternal Matrix Interface

# Train model
set.seed(123)
bolasso_mat <- bolasso(
  x = model.matrix(mpg ~ . - 1, mtcars_train),
  y = mtcars_train[, 1],
  data = mtcars_train,
  n.boot = 20,
  nfolds = 5,
  implement = "glmnet"
)

# Extract selected variables
selected_vars(bolasso_mat, threshold = 0.9, select = "lambda.min")
#> # A tibble: 2 × 2
#>   variable  mean_coef
#>   <chr>         <dbl>
#> 1 Intercept     21.5 
#> 2 wt            -3.04

# Bagged ensemble prediction on test data
predict(bolasso_mat,
        new.data = model.matrix(mpg ~ . - 1, mtcars_test),
        select = "lambda.min")
#>                       boot1     boot2    boot3    boot4    boot5    boot6
#> Valiant            20.63691 20.909550 19.82558 20.56410 19.73322 21.28793
#> Duster 360         15.76638 16.372454 14.02662 14.92655 15.52031 14.71217
#> Merc 230           25.12386 25.045148 22.92836 28.99311 21.73280 29.00448
#> Cadillac Fleetwood 11.87418  8.877457 12.36704 12.17847 12.54445 11.29728
#> Chrysler Imperial  11.09671  7.844000 11.18548 11.91890 11.61370 10.64259
#> Toyota Corolla     27.48744 29.725699 26.67354 31.37841 25.77683 31.20256
#> Toyota Corona      25.39340 26.295158 24.17597 30.16829 23.34638 27.62056
#> Fiat X1-9          26.47960 28.270351 26.43986 31.20191 25.50514 29.56025
#> Lotus Europa       26.33612 28.843855 25.38238 31.75100 25.21916 28.17203
#> Ferrari Dino       19.44345 20.549375 18.36977 21.48808 19.45716 19.70001
#>                       boot7    boot8    boot9   boot10   boot11    boot12
#> Valiant            20.88968 16.45979 21.81642 20.31133 18.31833 19.069060
#> Duster 360         16.12372 11.79266 14.05474 16.72145 16.61016 12.817112
#> Merc 230           25.87340 24.43973 31.50252 21.73801 21.90245 25.168089
#> Cadillac Fleetwood 11.68343 11.25757 11.60613 11.56359 10.84305  7.246436
#> Chrysler Imperial  10.71432 11.95087 11.02088 10.65026 10.32888  8.536854
#> Toyota Corolla     29.79012 27.14822 31.23569 27.17332 27.07122 24.570639
#> Toyota Corona      27.18482 20.52158 27.56142 24.16737 24.30330 26.303559
#> Fiat X1-9          28.79208 26.67265 28.51022 26.78921 26.65702 23.612131
#> Lotus Europa       28.00866 23.59150 25.15139 27.24218 27.61397 24.039371
#> Ferrari Dino       19.62817 20.66017 17.55627 21.20814 20.35387 20.173759
#>                      boot13   boot14    boot15    boot16   boot17   boot18
#> Valiant            20.96329 20.61230 22.217240 19.956577 18.65454 20.70182
#> Duster 360         11.75218 14.71489 14.937807 14.862105 16.51094 15.13084
#> Merc 230           29.39846 21.07515 25.881144 23.071813 21.66453 22.66934
#> Cadillac Fleetwood 14.80662 11.04649  7.117202 10.712186 11.63846 12.33739
#> Chrysler Imperial  14.51778 10.83905  5.906395  9.706662 11.01416 11.14948
#> Toyota Corolla     29.30298 23.94654 32.161672 26.195683 26.26276 27.17213
#> Toyota Corona      23.81786 22.57089 28.447409 23.310494 23.83429 24.34895
#> Fiat X1-9          27.27109 23.72819 30.423829 25.853065 25.92948 26.87812
#> Lotus Europa       22.72957 25.35814 29.083474 27.708635 26.68066 26.18377
#> Ferrari Dino       18.76525 24.25682 21.094349 20.146225 19.97039 19.82523
#>                      boot19    boot20
#> Valiant            19.74255 21.273184
#> Duster 360         16.04787 15.035304
#> Merc 230           23.13767 22.612717
#> Cadillac Fleetwood 12.57743 10.845459
#> Chrysler Imperial  11.85906  9.855632
#> Toyota Corolla     27.02833 27.582499
#> Toyota Corona      24.82676 24.713794
#> Fiat X1-9          26.75701 27.238644
#> Lotus Europa       26.92103 27.270005
#> Ferrari Dino       20.01838 21.511907