Bootsrap-enhanced Lasso

This function implements model-consistent Lasso estimation through the bootstrap. It supports parallel processing by way of the future package, allowing the user to flexibly specify many parallelization methods. This method was developed as a variable-selection algorithm, but this package also supports making ensemble predictions on new data using the bagged Lasso models.

Usage

bolasso(
  form,
  data,
  n.boot = 100,
  progress = TRUE,
  implement = "glmnet",
  x = NULL,
  y = NULL,
  ...
)

Arguments

form

An optional object of class formula (or one that can be coerced to that class): a symbolic description of the model to be fitted. Can be omitted when x and y are non-missing.

data

An optional object of class data.frame that contains the modeling variables referenced in form. Can be omitted when x and y are non-missing.

n.boot

An integer specifying the number of bootstrap replicates.

progress

A boolean indicating whether to display progress across bootstrap folds.

implement

A character; either 'glmnet' or 'gamlr', specifying which Lasso implementation to utilize. For specific modeling details, see glmnet::cv.glmnet or gamlr::cv.gamlr.

x

An optional predictor matrix in lieu of form and data.

y

An optional response vector in lieu of form and data.

...

Additional parameters to pass to either glmnet::cv.glmnet or gamlr::cv.gamlr.

Value

An object of class bolasso. This object is a list of length n.boot of cv.glmnet or cv.gamlr objects.

References

Bach FR (2008). “Bolasso: model consistent Lasso estimation through the bootstrap.” CoRR, abs/0804.1302. 0804.1302, http://arxiv.org/abs/0804.1302.

See also

Examples

mtcars[, c(2, 10:11)] <- lapply(mtcars[, c(2, 10:11)], as.factor)
idx <- sample(nrow(mtcars), 22)
mtcars_train <- mtcars[idx, ]
mtcars_test <- mtcars[-idx, ]

## Formula Interface

# Train model
set.seed(123)
bolasso_form <- bolasso(
  form = mpg ~ .,
  data = mtcars_train,
  n.boot = 20,
  nfolds = 5,
  implement = "glmnet"
)
#> Loaded glmnet 4.1-3

# Extract selected variables
selected_vars(bolasso_form, threshold = 0.9, select = "lambda.min")
#> # A tibble: 1 × 2
#>   variable  mean_coef
#>   <chr>         <dbl>
#> 1 Intercept      19.4

# Bagged ensemble prediction on test data
predict(bolasso_form,
        new.data = mtcars_test,
        select = "lambda.min")
#>                       boot1    boot2    boot3    boot4    boot5    boot6
#> Datsun 710         25.65541 27.71719 27.91953 25.42129 26.60702 26.24765
#> Valiant            19.31886 19.25259 19.62832 18.16273 18.95237 18.46365
#> Merc 450SL         16.93349 16.22702 16.51487 16.30560 15.99039 16.78384
#> Cadillac Fleetwood 15.25452 10.49457  9.20809 14.48676 11.15916 10.84713
#> Fiat 128           26.43830 28.50732 28.77371 26.10899 26.96336 26.91782
#> Honda Civic        26.87886 30.74136 30.49723 26.90419 31.45783 28.76560
#> Dodge Challenger   17.83798 17.34582 16.74411 17.13403 15.41781 17.74849
#> Pontiac Firebird   17.11308 15.85502 15.42925 16.31782 18.38423 16.51413
#> Lotus Europa       24.14958 30.28063 28.02049 24.28187 21.99687 26.41043
#> Volvo 142E         25.22609 25.83736 26.46071 24.68749 28.13320 24.73446
#>                       boot7    boot8     boot9   boot10    boot11   boot12
#> Datsun 710         24.61039 28.64636 26.635321 25.07470 27.509441 26.37090
#> Valiant            20.25044 18.08213 16.288637 22.53741 20.451600 18.40653
#> Merc 450SL         16.18316 16.16138 15.420547 18.35019 15.166961 15.90008
#> Cadillac Fleetwood 12.73186 11.37235  6.940375 16.06822  6.611105 11.86962
#> Fiat 128           25.14222 29.28533 27.644836 29.55036 29.976760 26.60141
#> Honda Civic        25.05228 30.89287 31.903280 30.19602 32.923024 32.44291
#> Dodge Challenger   15.35337 16.62442 17.716426 16.45509 14.149780 20.61464
#> Pontiac Firebird   14.03930 15.62940 16.239658 17.14884 13.112432 22.16396
#> Lotus Europa       22.48165 29.53410 27.116031 22.04963 28.871233 22.13981
#> Volvo 142E         23.62505 27.16040 26.023514 23.93835 24.736461 27.81858
#>                       boot13   boot14   boot15   boot16   boot17   boot18
#> Datsun 710         32.879890 25.39206 25.20000 30.37092 26.68028 25.66602
#> Valiant            10.163944 20.03394 20.11671 20.52895 20.54104 21.06293
#> Merc 450SL         15.742615 14.85100 15.74687 17.52370 15.63273 16.96151
#> Cadillac Fleetwood 11.309608 13.16988 11.33579 11.95542 11.23321 14.98832
#> Fiat 128           33.131798 26.88801 25.81619 31.36182 27.90918 26.47960
#> Honda Civic        29.808314 29.59474 27.13959 32.27397 29.32687 27.35410
#> Dodge Challenger   10.600513 14.65376 16.57876 18.11219 15.48427 17.93187
#> Pontiac Firebird    9.154091 14.67701 15.59748 16.75798 14.81197 16.99336
#> Lotus Europa       20.886887 26.64650 26.46393 29.79577 27.63564 25.57967
#> Volvo 142E         29.014826 25.55282 24.08806 28.32066 25.39952 24.84718
#>                      boot19   boot20
#> Datsun 710         26.27816 25.31357
#> Valiant            17.47877 19.43707
#> Merc 450SL         15.64247 16.84709
#> Cadillac Fleetwood 10.17352 13.64145
#> Fiat 128           26.29546 25.81236
#> Honda Civic        28.04138 27.08909
#> Dodge Challenger   16.71929 17.55060
#> Pontiac Firebird   16.24189 16.67222
#> Lotus Europa       24.80075 26.68884
#> Volvo 142E         25.13961 24.26064

## Alternal Matrix Interface

# Train model
set.seed(123)
bolasso_mat <- bolasso(
  x = model.matrix(mpg ~ . - 1, mtcars_train),
  y = mtcars_train[, 1],
  data = mtcars_train,
  n.boot = 20,
  nfolds = 5,
  implement = "glmnet"
)

# Extract selected variables
selected_vars(bolasso_mat, threshold = 0.9, select = "lambda.min")
#> # A tibble: 1 × 2
#>   variable  mean_coef
#>   <chr>         <dbl>
#> 1 Intercept      19.4

# Bagged ensemble prediction on test data
predict(bolasso_mat,
        new.data = model.matrix(mpg ~ . - 1, mtcars_test),
        select = "lambda.min")
#>                       boot1    boot2    boot3    boot4    boot5    boot6
#> Datsun 710         25.65541 27.71719 27.91953 25.42129 26.60702 26.24765
#> Valiant            19.31886 19.25259 19.62832 18.16273 18.95237 18.46365
#> Merc 450SL         16.93349 16.22702 16.51487 16.30560 15.99039 16.78384
#> Cadillac Fleetwood 15.25452 10.49457  9.20809 14.48676 11.15916 10.84713
#> Fiat 128           26.43830 28.50732 28.77371 26.10899 26.96336 26.91782
#> Honda Civic        26.87886 30.74136 30.49723 26.90419 31.45783 28.76560
#> Dodge Challenger   17.83798 17.34582 16.74411 17.13403 15.41781 17.74849
#> Pontiac Firebird   17.11308 15.85502 15.42925 16.31782 18.38423 16.51413
#> Lotus Europa       24.14958 30.28063 28.02049 24.28187 21.99687 26.41043
#> Volvo 142E         25.22609 25.83736 26.46071 24.68749 28.13320 24.73446
#>                       boot7    boot8     boot9   boot10    boot11   boot12
#> Datsun 710         24.61039 28.64636 26.635321 25.07470 27.509441 26.37090
#> Valiant            20.25044 18.08213 16.288637 22.53741 20.451600 18.40653
#> Merc 450SL         16.18316 16.16138 15.420547 18.35019 15.166961 15.90008
#> Cadillac Fleetwood 12.73186 11.37235  6.940375 16.06822  6.611105 11.86962
#> Fiat 128           25.14222 29.28533 27.644836 29.55036 29.976760 26.60141
#> Honda Civic        25.05228 30.89287 31.903280 30.19602 32.923024 32.44291
#> Dodge Challenger   15.35337 16.62442 17.716426 16.45509 14.149780 20.61464
#> Pontiac Firebird   14.03930 15.62940 16.239658 17.14884 13.112432 22.16396
#> Lotus Europa       22.48165 29.53410 27.116031 22.04963 28.871233 22.13981
#> Volvo 142E         23.62505 27.16040 26.023514 23.93835 24.736461 27.81858
#>                       boot13   boot14   boot15   boot16   boot17   boot18
#> Datsun 710         32.879890 25.39206 25.20000 30.37092 26.68028 25.66602
#> Valiant            10.163944 20.03394 20.11671 20.52895 20.54104 21.06293
#> Merc 450SL         15.742615 14.85100 15.74687 17.52370 15.63273 16.96151
#> Cadillac Fleetwood 11.309608 13.16988 11.33579 11.95542 11.23321 14.98832
#> Fiat 128           33.131798 26.88801 25.81619 31.36182 27.90918 26.47960
#> Honda Civic        29.808314 29.59474 27.13959 32.27397 29.32687 27.35410
#> Dodge Challenger   10.600513 14.65376 16.57876 18.11219 15.48427 17.93187
#> Pontiac Firebird    9.154091 14.67701 15.59748 16.75798 14.81197 16.99336
#> Lotus Europa       20.886887 26.64650 26.46393 29.79577 27.63564 25.57967
#> Volvo 142E         29.014826 25.55282 24.08806 28.32066 25.39952 24.84718
#>                      boot19   boot20
#> Datsun 710         26.27816 25.31357
#> Valiant            17.47877 19.43707
#> Merc 450SL         15.64247 16.84709
#> Cadillac Fleetwood 10.17352 13.64145
#> Fiat 128           26.29546 25.81236
#> Honda Civic        28.04138 27.08909
#> Dodge Challenger   16.71929 17.55060
#> Pontiac Firebird   16.24189 16.67222
#> Lotus Europa       24.80075 26.68884
#> Volvo 142E         25.13961 24.26064