Bootsrap-enhanced Lasso

This function implements model-consistent Lasso estimation through the bootstrap. It supports parallel processing by way of the future package, allowing the user to flexibly specify many parallelization methods. This method was developed as a variable-selection algorithm, but this package also supports making ensemble predictions on new data using the bagged Lasso models.

Usage

bolasso(
  form,
  data,
  n.boot = 100,
  progress = TRUE,
  implement = "glmnet",
  x = NULL,
  y = NULL,
  ...
)

Arguments

form

An optional object of class formula (or one that can be coerced to that class): a symbolic description of the model to be fitted. Can be omitted when x and y are non-missing.

data

An optional object of class data.frame that contains the modeling variables referenced in form. Can be omitted when x and y are non-missing.

n.boot

An integer specifying the number of bootstrap replicates.

progress

A boolean indicating whether to display progress across bootstrap folds.

implement

A character; either 'glmnet' or 'gamlr', specifying which Lasso implementation to utilize. For specific modeling details, see glmnet::cv.glmnet or gamlr::cv.gamlr.

x

An optional predictor matrix in lieu of form and data.

y

An optional response vector in lieu of form and data.

...

Additional parameters to pass to either glmnet::cv.glmnet or gamlr::cv.gamlr.

Value

An object of class bolasso. This object is a list of length n.boot of cv.glmnet or cv.gamlr objects.

References

Bach FR (2008). “Bolasso: model consistent Lasso estimation through the bootstrap.” CoRR, abs/0804.1302. 0804.1302, https://arxiv.org/abs/0804.1302.

See also

Examples

mtcars[, c(2, 10:11)] <- lapply(mtcars[, c(2, 10:11)], as.factor)
idx <- sample(nrow(mtcars), 22)
mtcars_train <- mtcars[idx, ]
mtcars_test <- mtcars[-idx, ]

## Formula Interface

# Train model
set.seed(123)
bolasso_form <- bolasso(
  form = mpg ~ .,
  data = mtcars_train,
  n.boot = 20,
  nfolds = 5,
  implement = "glmnet"
)
#> Loaded glmnet 4.1-3

# Extract selected variables
selected_vars(bolasso_form, threshold = 0.9, select = "lambda.min")
#> # A tibble: 2 × 2
#>   variable  mean_coef
#>   <chr>         <dbl>
#> 1 Intercept      1.64
#> 2 wt            -2.03

# Bagged ensemble prediction on test data
predict(bolasso_form,
        new.data = mtcars_test,
        select = "lambda.min")
#>                      boot1    boot2    boot3    boot4    boot5    boot6
#> Mazda RX4         27.30108 19.55748 18.86818 27.33368 20.01937 22.55160
#> Mazda RX4 Wag     28.84407 19.87669 19.54519 26.90695 20.35093 22.48992
#> Hornet 4 Drive    18.04460 17.07451 21.21822 18.75059 18.15604 18.05761
#> Valiant           19.42097 16.71314 19.79784 18.63613 17.43208 18.17011
#> Merc 230          30.93987 29.33743 30.14795 27.87086 28.32019 26.84653
#> Merc 450SL        16.09559 14.94398 16.23420 16.40156 15.96032 17.08967
#> Chrysler Imperial 16.84525 10.70768 12.11024 11.20229 12.28222 11.36773
#> Dodge Challenger  18.44248 13.15442 15.61173 16.59254 14.21728 17.03976
#> Lotus Europa      18.53760 24.93065 21.60473 22.89581 21.24979 27.06231
#> Volvo 142E        24.80737 26.04855 29.18683 28.19066 25.80991 28.18690
#>                       boot7    boot8    boot9   boot10   boot11   boot12
#> Mazda RX4          7.173597 24.37747 25.68187 20.34458 21.67765 19.95889
#> Mazda RX4 Wag     10.135256 23.84963 25.12290 20.00407 21.52916 19.71204
#> Hornet 4 Drive    16.167026 19.08213 20.93182 20.13402 19.90093 20.24742
#> Valiant            9.966407 18.47625 20.61317 19.01512 19.54792 19.58109
#> Merc 230          46.444186 21.83329 21.97090 24.14943 27.10019 26.65045
#> Merc 450SL        14.317964 15.71427 16.98680 17.66261 15.88451 15.70857
#> Chrysler Imperial 16.497036 10.86427 11.26272 10.89140 10.55814 10.67613
#> Dodge Challenger   6.129797 16.97831 18.75750 17.77744 15.68901 16.78866
#> Lotus Europa      10.970837 25.52578 28.21882 24.56865 27.02121 26.33749
#> Volvo 142E        36.330887 24.50021 25.61623 24.65023 25.47887 25.01095
#>                     boot13    boot14   boot15   boot16   boot17   boot18
#> Mazda RX4         21.07500 21.973744 21.87521 22.25342 21.96200 22.21354
#> Mazda RX4 Wag     20.80065 20.821554 21.50177 21.74571 21.71778 21.45817
#> Hornet 4 Drive    20.58222 20.349787 21.54436 17.08875 19.60526 21.13260
#> Valiant           20.55141 18.690524 21.07824 16.65019 19.06443 20.53512
#> Merc 230          27.76152 21.654221 26.21864 24.63166 25.76915 22.18598
#> Merc 450SL        17.16985 16.192894 16.87999 15.37412 15.68601 17.81127
#> Chrysler Imperial 10.40530  6.047878 10.40175 10.86645 10.31478 10.57517
#> Dodge Challenger  16.21480 16.865940 17.07459 16.08763 15.16975 19.67899
#> Lotus Europa      28.85082 27.452592 25.94803 25.67744 26.52092 27.06121
#> Volvo 142E        27.70860 23.174039 23.62983 26.60231 24.80764 23.41070
#>                     boot19   boot20
#> Mazda RX4         24.08124 23.20341
#> Mazda RX4 Wag     24.17197 23.23430
#> Hornet 4 Drive    20.20774 19.53153
#> Valiant           20.63196 19.89219
#> Merc 230          24.23999 27.66367
#> Merc 450SL        16.18159 15.17728
#> Chrysler Imperial 13.77349 10.33199
#> Dodge Challenger  18.90784 17.33990
#> Lotus Europa      25.69091 27.27966
#> Volvo 142E        27.33689 27.57869

## Alternal Matrix Interface

# Train model
set.seed(123)
bolasso_mat <- bolasso(
  x = model.matrix(mpg ~ . - 1, mtcars_train),
  y = mtcars_train[, 1],
  data = mtcars_train,
  n.boot = 20,
  nfolds = 5,
  implement = "glmnet"
)

# Extract selected variables
selected_vars(bolasso_mat, threshold = 0.9, select = "lambda.min")
#> # A tibble: 2 × 2
#>   variable  mean_coef
#>   <chr>         <dbl>
#> 1 Intercept      1.64
#> 2 wt            -2.03

# Bagged ensemble prediction on test data
predict(bolasso_mat,
        new.data = model.matrix(mpg ~ . - 1, mtcars_test),
        select = "lambda.min")
#>                      boot1    boot2    boot3    boot4    boot5    boot6
#> Mazda RX4         27.30108 19.55748 18.86818 27.33368 20.01937 22.55160
#> Mazda RX4 Wag     28.84407 19.87669 19.54519 26.90695 20.35093 22.48992
#> Hornet 4 Drive    18.04460 17.07451 21.21822 18.75059 18.15604 18.05761
#> Valiant           19.42097 16.71314 19.79784 18.63613 17.43208 18.17011
#> Merc 230          30.93987 29.33743 30.14795 27.87086 28.32019 26.84653
#> Merc 450SL        16.09559 14.94398 16.23420 16.40156 15.96032 17.08967
#> Chrysler Imperial 16.84525 10.70768 12.11024 11.20229 12.28222 11.36773
#> Dodge Challenger  18.44248 13.15442 15.61173 16.59254 14.21728 17.03976
#> Lotus Europa      18.53760 24.93065 21.60473 22.89581 21.24979 27.06231
#> Volvo 142E        24.80737 26.04855 29.18683 28.19066 25.80991 28.18690
#>                       boot7    boot8    boot9   boot10   boot11   boot12
#> Mazda RX4          7.173597 24.37747 25.68187 20.34458 21.67765 19.95889
#> Mazda RX4 Wag     10.135256 23.84963 25.12290 20.00407 21.52916 19.71204
#> Hornet 4 Drive    16.167026 19.08213 20.93182 20.13402 19.90093 20.24742
#> Valiant            9.966407 18.47625 20.61317 19.01512 19.54792 19.58109
#> Merc 230          46.444186 21.83329 21.97090 24.14943 27.10019 26.65045
#> Merc 450SL        14.317964 15.71427 16.98680 17.66261 15.88451 15.70857
#> Chrysler Imperial 16.497036 10.86427 11.26272 10.89140 10.55814 10.67613
#> Dodge Challenger   6.129797 16.97831 18.75750 17.77744 15.68901 16.78866
#> Lotus Europa      10.970837 25.52578 28.21882 24.56865 27.02121 26.33749
#> Volvo 142E        36.330887 24.50021 25.61623 24.65023 25.47887 25.01095
#>                     boot13    boot14   boot15   boot16   boot17   boot18
#> Mazda RX4         21.07500 21.973744 21.87521 22.25342 21.96200 22.21354
#> Mazda RX4 Wag     20.80065 20.821554 21.50177 21.74571 21.71778 21.45817
#> Hornet 4 Drive    20.58222 20.349787 21.54436 17.08875 19.60526 21.13260
#> Valiant           20.55141 18.690524 21.07824 16.65019 19.06443 20.53512
#> Merc 230          27.76152 21.654221 26.21864 24.63166 25.76915 22.18598
#> Merc 450SL        17.16985 16.192894 16.87999 15.37412 15.68601 17.81127
#> Chrysler Imperial 10.40530  6.047878 10.40175 10.86645 10.31478 10.57517
#> Dodge Challenger  16.21480 16.865940 17.07459 16.08763 15.16975 19.67899
#> Lotus Europa      28.85082 27.452592 25.94803 25.67744 26.52092 27.06121
#> Volvo 142E        27.70860 23.174039 23.62983 26.60231 24.80764 23.41070
#>                     boot19   boot20
#> Mazda RX4         24.08124 23.20341
#> Mazda RX4 Wag     24.17197 23.23430
#> Hornet 4 Drive    20.20774 19.53153
#> Valiant           20.63196 19.89219
#> Merc 230          24.23999 27.66367
#> Merc 450SL        16.18159 15.17728
#> Chrysler Imperial 13.77349 10.33199
#> Dodge Challenger  18.90784 17.33990
#> Lotus Europa      25.69091 27.27966
#> Volvo 142E        27.33689 27.57869