Bootsrap-enhanced Lasso

This function implements model-consistent Lasso estimation through the bootstrap. It supports parallel processing by way of the future package, allowing the user to flexibly specify many parallelization methods. This method was developed as a variable-selection algorithm, but this package also supports making ensemble predictions on new data using the bagged Lasso models.

Usage

bolasso(
  form,
  data,
  n.boot = 100,
  progress = TRUE,
  implement = "glmnet",
  x = NULL,
  y = NULL,
  ...
)

Arguments

form

An optional object of class formula (or one that can be coerced to that class): a symbolic description of the model to be fitted. Can be omitted when x and y are non-missing.

data

An optional object of class data.frame that contains the modeling variables referenced in form. Can be omitted when x and y are non-missing.

n.boot

An integer specifying the number of bootstrap replicates.

progress

A boolean indicating whether to display progress across bootstrap folds.

implement

A character; either 'glmnet' or 'gamlr', specifying which Lasso implementation to utilize. For specific modeling details, see glmnet::cv.glmnet or gamlr::cv.gamlr.

x

An optional predictor matrix in lieu of form and data.

y

An optional response vector in lieu of form and data.

...

Additional parameters to pass to either glmnet::cv.glmnet or gamlr::cv.gamlr.

Value

An object of class bolasso. This object is a list of length n.boot of cv.glmnet or cv.gamlr objects.

References

Bach FR (2008). “Bolasso: model consistent Lasso estimation through the bootstrap.” CoRR, abs/0804.1302. 0804.1302, https://arxiv.org/abs/0804.1302.

See also

Examples

mtcars[, c(2, 10:11)] <- lapply(mtcars[, c(2, 10:11)], as.factor)
idx <- sample(nrow(mtcars), 22)
mtcars_train <- mtcars[idx, ]
mtcars_test <- mtcars[-idx, ]

## Formula Interface

# Train model
set.seed(123)
bolasso_form <- bolasso(
  form = mpg ~ .,
  data = mtcars_train,
  n.boot = 20,
  nfolds = 5,
  implement = "glmnet"
)
#> Loaded glmnet 4.1-3

# Extract selected variables
selected_vars(bolasso_form, threshold = 0.9, select = "lambda.min")
#> # A tibble: 3 × 2
#>   variable  mean_coef
#>   <chr>         <dbl>
#> 1 Intercept     25.4 
#> 2 cyl4           3.06
#> 3 wt            -1.93

# Bagged ensemble prediction on test data
predict(bolasso_form,
        new.data = mtcars_test,
        select = "lambda.min")
#>                      boot1    boot2    boot3    boot4    boot5    boot6
#> Datsun 710        25.43955 31.79118 24.76856 28.84549 30.01312 29.40558
#> Hornet 4 Drive    19.46857 20.68861 20.76448 19.14018 18.62320 19.28849
#> Hornet Sportabout 18.03213 13.71400 18.61155 18.02063 17.40989 17.53204
#> Chrysler Imperial 11.09702 10.23712 10.27052 10.20298 10.52686 10.64544
#> Toyota Corolla    27.85827 32.92310 26.35663 30.51187 31.46861 30.33498
#> AMC Javelin       18.51603 14.29219 16.94966 18.36212 17.68055 17.54162
#> Fiat X1-9         27.32203 33.09013 25.96641 30.23093 31.21789 30.14335
#> Lotus Europa      26.99630 28.95446 26.83018 29.35190 28.76497 28.50754
#> Ford Pantera L    17.58976 16.02875 18.30236 16.76378 15.27969 18.33395
#> Volvo 142E        24.75251 26.71531 25.14671 27.40355 28.74187 28.52407
#>                      boot7    boot8     boot9   boot10   boot11   boot12
#> Datsun 710        32.60356 26.30909 22.949099 18.18804 24.45882 26.32437
#> Hornet 4 Drive    18.99687 16.73668 18.954464 14.81463 20.28072 16.76563
#> Hornet Sportabout 17.79770 14.50419 20.469405 19.51540 19.90432 13.81210
#> Chrysler Imperial 16.13253 10.92954  9.634968 13.04671 10.24853 11.49218
#> Toyota Corolla    32.18455 27.60181 25.050766 20.37997 25.62819 27.39667
#> AMC Javelin       16.41241 15.48379 21.200912 18.93780 20.43576 15.42455
#> Fiat X1-9         32.08268 27.38518 24.915507 20.22988 25.77217 27.17135
#> Lotus Europa      26.87831 26.28352 28.552956 18.77866 29.35034 26.67389
#> Ford Pantera L    20.53487 15.89178 16.179892 17.60834 16.52071 13.99208
#> Volvo 142E        33.12509 25.50110 24.843357 21.71925 26.56969 25.94138
#>                     boot13   boot14    boot15   boot16   boot17   boot18
#> Datsun 710        32.76185 27.71787 26.101983 27.35805 31.61983 24.66154
#> Hornet 4 Drive    17.45322 19.40099 17.616329 17.03055 26.41753 19.40271
#> Hornet Sportabout 14.73634 18.58559 15.506500 15.31213 20.00002 17.78651
#> Chrysler Imperial 13.86020 11.86625  9.664321 11.62027 10.46767 11.52496
#> Toyota Corolla    32.03173 29.44409 27.853112 28.56457 32.93901 26.35038
#> AMC Javelin       14.83339 18.98979 16.347966 15.98471 20.45725 17.79017
#> Fiat X1-9         32.46222 29.16281 27.501737 28.36420 33.05582 25.77251
#> Lotus Europa      32.76097 30.34762 27.168788 27.02694 22.89214 24.91105
#> Ford Pantera L    23.60009 17.11246 16.026446 17.56039 10.33015 18.69990
#> Volvo 142E        31.55360 27.04439 24.769020 26.48310 26.30530 26.14969
#>                     boot19   boot20
#> Datsun 710        23.52573 22.22575
#> Hornet 4 Drive    18.86192 17.56090
#> Hornet Sportabout 20.22091 19.92278
#> Chrysler Imperial  9.94161 10.65696
#> Toyota Corolla    25.38748 23.28116
#> AMC Javelin       20.75751 20.85556
#> Fiat X1-9         25.10116 23.41179
#> Lotus Europa      27.38361 26.83518
#> Ford Pantera L    15.69026 15.44703
#> Volvo 142E        25.30705 25.14344

## Alternal Matrix Interface

# Train model
set.seed(123)
bolasso_mat <- bolasso(
  x = model.matrix(mpg ~ . - 1, mtcars_train),
  y = mtcars_train[, 1],
  data = mtcars_train,
  n.boot = 20,
  nfolds = 5,
  implement = "glmnet"
)

# Extract selected variables
selected_vars(bolasso_mat, threshold = 0.9, select = "lambda.min")
#> # A tibble: 3 × 2
#>   variable  mean_coef
#>   <chr>         <dbl>
#> 1 Intercept     25.4 
#> 2 cyl4           3.06
#> 3 wt            -1.93

# Bagged ensemble prediction on test data
predict(bolasso_mat,
        new.data = model.matrix(mpg ~ . - 1, mtcars_test),
        select = "lambda.min")
#>                      boot1    boot2    boot3    boot4    boot5    boot6
#> Datsun 710        25.43955 31.79118 24.76856 28.84549 30.01312 29.40558
#> Hornet 4 Drive    19.46857 20.68861 20.76448 19.14018 18.62320 19.28849
#> Hornet Sportabout 18.03213 13.71400 18.61155 18.02063 17.40989 17.53204
#> Chrysler Imperial 11.09702 10.23712 10.27052 10.20298 10.52686 10.64544
#> Toyota Corolla    27.85827 32.92310 26.35663 30.51187 31.46861 30.33498
#> AMC Javelin       18.51603 14.29219 16.94966 18.36212 17.68055 17.54162
#> Fiat X1-9         27.32203 33.09013 25.96641 30.23093 31.21789 30.14335
#> Lotus Europa      26.99630 28.95446 26.83018 29.35190 28.76497 28.50754
#> Ford Pantera L    17.58976 16.02875 18.30236 16.76378 15.27969 18.33395
#> Volvo 142E        24.75251 26.71531 25.14671 27.40355 28.74187 28.52407
#>                      boot7    boot8     boot9   boot10   boot11   boot12
#> Datsun 710        32.60356 26.30909 22.949099 18.18804 24.45882 26.32437
#> Hornet 4 Drive    18.99687 16.73668 18.954464 14.81463 20.28072 16.76563
#> Hornet Sportabout 17.79770 14.50419 20.469405 19.51540 19.90432 13.81210
#> Chrysler Imperial 16.13253 10.92954  9.634968 13.04671 10.24853 11.49218
#> Toyota Corolla    32.18455 27.60181 25.050766 20.37997 25.62819 27.39667
#> AMC Javelin       16.41241 15.48379 21.200912 18.93780 20.43576 15.42455
#> Fiat X1-9         32.08268 27.38518 24.915507 20.22988 25.77217 27.17135
#> Lotus Europa      26.87831 26.28352 28.552956 18.77866 29.35034 26.67389
#> Ford Pantera L    20.53487 15.89178 16.179892 17.60834 16.52071 13.99208
#> Volvo 142E        33.12509 25.50110 24.843357 21.71925 26.56969 25.94138
#>                     boot13   boot14    boot15   boot16   boot17   boot18
#> Datsun 710        32.76185 27.71787 26.101983 27.35805 31.61983 24.66154
#> Hornet 4 Drive    17.45322 19.40099 17.616329 17.03055 26.41753 19.40271
#> Hornet Sportabout 14.73634 18.58559 15.506500 15.31213 20.00002 17.78651
#> Chrysler Imperial 13.86020 11.86625  9.664321 11.62027 10.46767 11.52496
#> Toyota Corolla    32.03173 29.44409 27.853112 28.56457 32.93901 26.35038
#> AMC Javelin       14.83339 18.98979 16.347966 15.98471 20.45725 17.79017
#> Fiat X1-9         32.46222 29.16281 27.501737 28.36420 33.05582 25.77251
#> Lotus Europa      32.76097 30.34762 27.168788 27.02694 22.89214 24.91105
#> Ford Pantera L    23.60009 17.11246 16.026446 17.56039 10.33015 18.69990
#> Volvo 142E        31.55360 27.04439 24.769020 26.48310 26.30530 26.14969
#>                     boot19   boot20
#> Datsun 710        23.52573 22.22575
#> Hornet 4 Drive    18.86192 17.56090
#> Hornet Sportabout 20.22091 19.92278
#> Chrysler Imperial  9.94161 10.65696
#> Toyota Corolla    25.38748 23.28116
#> AMC Javelin       20.75751 20.85556
#> Fiat X1-9         25.10116 23.41179
#> Lotus Europa      27.38361 26.83518
#> Ford Pantera L    15.69026 15.44703
#> Volvo 142E        25.30705 25.14344