Bootsrap-enhanced Lasso

This function implements model-consistent Lasso estimation through the bootstrap. It supports parallel processing by way of the future package, allowing the user to flexibly specify many parallelization methods. This method was developed as a variable-selection algorithm, but this package also supports making ensemble predictions on new data using the bagged Lasso models.

Usage

bolasso(
  form,
  data,
  n.boot = 100,
  progress = TRUE,
  implement = "glmnet",
  x = NULL,
  y = NULL,
  ...
)

Arguments

form

An optional object of class formula (or one that can be coerced to that class): a symbolic description of the model to be fitted. Can be omitted when x and y are non-missing.

data

An optional object of class data.frame that contains the modeling variables referenced in form. Can be omitted when x and y are non-missing.

n.boot

An integer specifying the number of bootstrap replicates.

progress

A boolean indicating whether to display progress across bootstrap folds.

implement

A character; either 'glmnet' or 'gamlr', specifying which Lasso implementation to utilize. For specific modeling details, see glmnet::cv.glmnet or gamlr::cv.gamlr.

x

An optional predictor matrix in lieu of form and data.

y

An optional response vector in lieu of form and data.

...

Additional parameters to pass to either glmnet::cv.glmnet or gamlr::cv.gamlr.

Value

An object of class bolasso. This object is a list of length n.boot of cv.glmnet or cv.gamlr objects.

References

Bach FR (2008). “Bolasso: model consistent Lasso estimation through the bootstrap.” CoRR, abs/0804.1302. 0804.1302, https://arxiv.org/abs/0804.1302.

See also

Examples

mtcars[, c(2, 10:11)] <- lapply(mtcars[, c(2, 10:11)], as.factor)
idx <- sample(nrow(mtcars), 22)
mtcars_train <- mtcars[idx, ]
mtcars_test <- mtcars[-idx, ]

## Formula Interface

# Train model
set.seed(123)
bolasso_form <- bolasso(
  form = mpg ~ .,
  data = mtcars_train,
  n.boot = 20,
  nfolds = 5,
  implement = "glmnet"
)
#> Loaded glmnet 4.1-3

# Extract selected variables
selected_vars(bolasso_form, threshold = 0.9, select = "lambda.min")
#> # A tibble: 1 × 2
#>   variable  mean_coef
#>   <chr>         <dbl>
#> 1 Intercept      12.3

# Bagged ensemble prediction on test data
predict(bolasso_form,
        new.data = mtcars_test,
        select = "lambda.min")
#>                         boot1    boot2    boot3    boot4     boot5    boot6
#> Mazda RX4           21.576581 17.72208 17.81363 21.20254 22.531935 21.66663
#> Hornet 4 Drive      21.315313 18.88140 18.14156 20.63045 20.231857 19.50443
#> Merc 230            28.563419 27.53128 31.72019 27.71459 23.639463 24.95103
#> Merc 450SE          14.634207 15.37023 15.94594 17.98082 13.462257 16.69345
#> Cadillac Fleetwood  10.593666 14.77655 14.65380 16.38941  6.272891 15.38237
#> Lincoln Continental  9.646374 14.63065 14.26808 16.01089  5.104668 15.05125
#> Toyota Corona       28.143758 24.22182 27.15060 27.63889 25.962142 25.35175
#> Dodge Challenger    16.828665 19.27542 16.07143 19.11638 16.423693 17.70497
#> Ford Pantera L      16.868116 16.00834 18.23128 15.37336 17.943056 17.94893
#> Volvo 142E          26.415154 28.79551 26.84974 28.40188 23.277864 27.39038
#>                         boot7    boot8    boot9   boot10   boot11    boot12
#> Mazda RX4           22.071337 21.80766 20.64976 18.54756 16.25218 19.931657
#> Hornet 4 Drive      22.059909 18.68733 21.20893 18.50071 26.14215 19.040273
#> Merc 230            28.359383 24.17637 29.27342 34.61890 34.70485 23.914134
#> Merc 450SE          14.811099 16.35465 14.44790 14.27509 14.38116 15.690701
#> Cadillac Fleetwood  10.553870 15.37268 15.42746 14.95494 15.19372  8.458725
#> Lincoln Continental  9.454356 15.03062 15.08545 14.07267 14.71797  7.654925
#> Toyota Corona       28.370220 22.93583 27.78157 27.88506 30.97364 23.103285
#> Dodge Challenger    18.172844 18.68024 15.02360 12.72952 16.15735 20.837440
#> Ford Pantera L      17.660721 18.08754 18.08921 22.33185 17.80064 15.916392
#> Volvo 142E          25.583624 26.82097 30.70053 23.81754 25.79784 25.198514
#>                       boot13    boot14    boot15   boot16   boot17   boot18
#> Mazda RX4           21.67093 21.038541 22.006916 20.90707 21.29648 21.30790
#> Hornet 4 Drive      18.22302 22.701018 19.454005 19.52332 20.27347 20.18572
#> Merc 230            24.29636 28.644075 23.388299 23.69538 24.30045 29.27176
#> Merc 450SE          16.32147 15.455269 14.792445 17.05751 16.97800 15.47429
#> Cadillac Fleetwood  15.79774  8.355012 10.048207 15.17764 13.82099 14.42646
#> Lincoln Continental 15.59320  7.197733  9.318878 14.86097 13.28489 13.64173
#> Toyota Corona       23.48012 27.966383 24.985144 21.56781 25.94852 26.20196
#> Dodge Challenger    16.94068 18.352384 17.917855 16.43660 18.65459 18.84163
#> Ford Pantera L      17.70188 15.680787 17.756155 17.12650 18.23368 19.66614
#> Volvo 142E          27.05537 25.100132 24.708139 25.84248 25.04622 27.69798
#>                       boot19   boot20
#> Mazda RX4           20.95239 20.92704
#> Hornet 4 Drive      19.50658 19.41958
#> Merc 230            29.06323 26.28186
#> Merc 450SE          16.47108 15.81907
#> Cadillac Fleetwood  14.58992 13.18902
#> Lincoln Continental 14.42696 12.63278
#> Toyota Corona       26.23110 24.77776
#> Dodge Challenger    18.51523 18.80185
#> Ford Pantera L      17.81777 17.15225
#> Volvo 142E          30.91216 26.45386

## Alternal Matrix Interface

# Train model
set.seed(123)
bolasso_mat <- bolasso(
  x = model.matrix(mpg ~ . - 1, mtcars_train),
  y = mtcars_train[, 1],
  data = mtcars_train,
  n.boot = 20,
  nfolds = 5,
  implement = "glmnet"
)

# Extract selected variables
selected_vars(bolasso_mat, threshold = 0.9, select = "lambda.min")
#> # A tibble: 1 × 2
#>   variable  mean_coef
#>   <chr>         <dbl>
#> 1 Intercept      12.3

# Bagged ensemble prediction on test data
predict(bolasso_mat,
        new.data = model.matrix(mpg ~ . - 1, mtcars_test),
        select = "lambda.min")
#>                         boot1    boot2    boot3    boot4     boot5    boot6
#> Mazda RX4           21.576581 17.72208 17.81363 21.20254 22.531935 21.66663
#> Hornet 4 Drive      21.315313 18.88140 18.14156 20.63045 20.231857 19.50443
#> Merc 230            28.563419 27.53128 31.72019 27.71459 23.639463 24.95103
#> Merc 450SE          14.634207 15.37023 15.94594 17.98082 13.462257 16.69345
#> Cadillac Fleetwood  10.593666 14.77655 14.65380 16.38941  6.272891 15.38237
#> Lincoln Continental  9.646374 14.63065 14.26808 16.01089  5.104668 15.05125
#> Toyota Corona       28.143758 24.22182 27.15060 27.63889 25.962142 25.35175
#> Dodge Challenger    16.828665 19.27542 16.07143 19.11638 16.423693 17.70497
#> Ford Pantera L      16.868116 16.00834 18.23128 15.37336 17.943056 17.94893
#> Volvo 142E          26.415154 28.79551 26.84974 28.40188 23.277864 27.39038
#>                         boot7    boot8    boot9   boot10   boot11    boot12
#> Mazda RX4           22.071337 21.80766 20.64976 18.54756 16.25218 19.931657
#> Hornet 4 Drive      22.059909 18.68733 21.20893 18.50071 26.14215 19.040273
#> Merc 230            28.359383 24.17637 29.27342 34.61890 34.70485 23.914134
#> Merc 450SE          14.811099 16.35465 14.44790 14.27509 14.38116 15.690701
#> Cadillac Fleetwood  10.553870 15.37268 15.42746 14.95494 15.19372  8.458725
#> Lincoln Continental  9.454356 15.03062 15.08545 14.07267 14.71797  7.654925
#> Toyota Corona       28.370220 22.93583 27.78157 27.88506 30.97364 23.103285
#> Dodge Challenger    18.172844 18.68024 15.02360 12.72952 16.15735 20.837440
#> Ford Pantera L      17.660721 18.08754 18.08921 22.33185 17.80064 15.916392
#> Volvo 142E          25.583624 26.82097 30.70053 23.81754 25.79784 25.198514
#>                       boot13    boot14    boot15   boot16   boot17   boot18
#> Mazda RX4           21.67093 21.038541 22.006916 20.90707 21.29648 21.30790
#> Hornet 4 Drive      18.22302 22.701018 19.454005 19.52332 20.27347 20.18572
#> Merc 230            24.29636 28.644075 23.388299 23.69538 24.30045 29.27176
#> Merc 450SE          16.32147 15.455269 14.792445 17.05751 16.97800 15.47429
#> Cadillac Fleetwood  15.79774  8.355012 10.048207 15.17764 13.82099 14.42646
#> Lincoln Continental 15.59320  7.197733  9.318878 14.86097 13.28489 13.64173
#> Toyota Corona       23.48012 27.966383 24.985144 21.56781 25.94852 26.20196
#> Dodge Challenger    16.94068 18.352384 17.917855 16.43660 18.65459 18.84163
#> Ford Pantera L      17.70188 15.680787 17.756155 17.12650 18.23368 19.66614
#> Volvo 142E          27.05537 25.100132 24.708139 25.84248 25.04622 27.69798
#>                       boot19   boot20
#> Mazda RX4           20.95239 20.92704
#> Hornet 4 Drive      19.50658 19.41958
#> Merc 230            29.06323 26.28186
#> Merc 450SE          16.47108 15.81907
#> Cadillac Fleetwood  14.58992 13.18902
#> Lincoln Continental 14.42696 12.63278
#> Toyota Corona       26.23110 24.77776
#> Dodge Challenger    18.51523 18.80185
#> Ford Pantera L      17.81777 17.15225
#> Volvo 142E          30.91216 26.45386