Bootsrap-enhanced Lasso

This function implements model-consistent Lasso estimation through the bootstrap. It supports parallel processing by way of the future package, allowing the user to flexibly specify many parallelization methods. This method was developed as a variable-selection algorithm, but this package also supports making ensemble predictions on new data using the bagged Lasso models.

Usage

bolasso(
  formula,
  data,
  n.boot = 100,
  progress = TRUE,
  implement = "glmnet",
  x = NULL,
  y = NULL,
  ...
)

Arguments

formula

An optional object of class formula (or one that can be coerced to that class): a symbolic description of the model to be fitted. Can be omitted when x and y are non-missing.

data

An optional object of class data.frame that contains the modeling variables referenced in form. Can be omitted when x and y are non-missing.

n.boot

An integer specifying the number of bootstrap replicates.

progress

A boolean indicating whether to display progress across bootstrap folds.

implement

A character; either 'glmnet' or 'gamlr', specifying which Lasso implementation to utilize. For specific modeling details, see glmnet::cv.glmnet or gamlr::cv.gamlr.

x

An optional predictor matrix in lieu of form and data.

y

An optional response vector in lieu of form and data.

...

Additional parameters to pass to either glmnet::cv.glmnet or gamlr::cv.gamlr.

Value

An object of class bolasso. This object is a list of length n.boot of cv.glmnet or cv.gamlr objects.

References

Bach FR (2008). “Bolasso: model consistent Lasso estimation through the bootstrap.” CoRR, abs/0804.1302. 0804.1302, https://arxiv.org/abs/0804.1302.

See also

Examples

mtcars[, c(2, 10:11)] <- lapply(mtcars[, c(2, 10:11)], as.factor)
idx <- sample(nrow(mtcars), 22)
mtcars_train <- mtcars[idx, ]
mtcars_test <- mtcars[-idx, ]

## Formula Interface

# Train model
set.seed(123)
bolasso_form <- bolasso(
  form = mpg ~ .,
  data = mtcars_train,
  n.boot = 20,
  nfolds = 5,
  implement = "glmnet"
)
#> Loaded glmnet 4.1-4

# Extract selected variables
selected_vars(bolasso_form, threshold = 0.9, select = "lambda.min")
#> # A tibble: 2 × 2
#>   variable  mean_coef
#>   <chr>         <dbl>
#> 1 Intercept     10.9 
#> 2 drat           3.10

# Bagged ensemble prediction on test data
predict(bolasso_form,
        new.data = mtcars_test,
        select = "lambda.min")
#>                   boot1    boot2    boot3    boot4    boot5    boot6    boot7
#> Datsun 710     28.23739 22.76799 26.14163 30.79058 24.39929 26.99998 27.26240
#> Duster 360     13.82016 13.66304 16.80029 13.57099 16.12820 16.36782 17.41064
#> Merc 230       23.42902 23.03986 25.86909 26.32375 21.44036 27.21191 25.16401
#> Merc 280C      20.14410 19.69072 20.90703 17.55778 19.75631 20.14694 19.83804
#> Merc 450SLC    16.70856 16.89269 17.91436 16.95218 16.84538 16.80853 17.08464
#> Fiat 128       29.25628 24.43285 28.02294 31.80075 25.55741 28.05651 28.05296
#> Camaro Z28     13.16418 16.02415 16.37099 16.54417 15.39750 16.23831 16.48265
#> Fiat X1-9      29.90018 24.43285 28.49873 31.80075 26.49691 28.38153 28.93802
#> Lotus Europa   26.93777 21.94509 25.34544 23.40932 26.75696 27.01956 29.10974
#> Ford Pantera L 17.01820 17.81238 18.64044 20.76708 17.54120 18.20846 19.07608
#>                   boot8    boot9   boot10   boot11   boot12   boot13   boot14
#> Datsun 710     25.05114 28.32561 23.95276 26.07585 22.46841 19.18533 29.87324
#> Duster 360     18.11097 16.31502 16.11316 17.12337 10.34018 13.72910 15.09769
#> Merc 230       23.77558 31.24133 26.29606 29.53528 22.33827 25.65757 25.37163
#> Merc 280C      20.14600 20.44868 21.57800 21.48436 19.27682 25.50253 19.44269
#> Merc 450SLC    17.46243 18.37710 17.30952 18.23548 14.03156 16.25785 18.05621
#> Fiat 128       25.61771 30.91778 25.79213 28.39177 24.66792 21.02080 30.44192
#> Camaro Z28     18.19168 16.95637 16.52547 16.33836 12.05943 18.37287 12.43247
#> Fiat X1-9      26.20529 30.69151 25.90452 28.13491 25.00436 21.13954 31.59828
#> Lotus Europa   27.20930 25.25403 24.08508 23.12581 23.54140 18.38618 18.85396
#> Ford Pantera L 20.31744 19.76050 18.78042 17.44687 15.82442 16.78050 15.94418
#>                  boot15   boot16   boot17   boot18   boot19   boot20
#> Datsun 710     23.79072 18.70437 28.92896 25.02703 30.06894 25.37033
#> Duster 360     14.39689 13.70505 18.16477 15.62079 17.36808 14.42852
#> Merc 230       22.53595 20.86224 23.28615 25.46959 26.27773 21.53995
#> Merc 280C      19.45909 24.91497 19.33958 20.29652 18.98660 21.21375
#> Merc 450SLC    17.13675 16.35214 17.62744 16.52344 17.66601 16.43115
#> Fiat 128       25.21987 21.18300 29.74118 26.65879 31.10353 27.11422
#> Camaro Z28     14.69220 19.30893 17.72204 15.67979 15.85902 15.14586
#> Fiat X1-9      25.76899 21.18300 30.57515 26.95144 32.04494 27.52984
#> Lotus Europa   24.90745 21.05049 27.18999 25.75213 29.52043 24.55380
#> Ford Pantera L 16.08225 23.18600 19.91582 18.97951 18.51803 17.76267

## Alternal Matrix Interface

# Train model
set.seed(123)
bolasso_mat <- bolasso(
  x = model.matrix(mpg ~ . - 1, mtcars_train),
  y = mtcars_train[, 1],
  data = mtcars_train,
  n.boot = 20,
  nfolds = 5,
  implement = "glmnet"
)

# Extract selected variables
selected_vars(bolasso_mat, threshold = 0.9, select = "lambda.min")
#> # A tibble: 2 × 2
#>   variable  mean_coef
#>   <chr>         <dbl>
#> 1 Intercept     10.9 
#> 2 drat           3.10

# Bagged ensemble prediction on test data
predict(bolasso_mat,
        new.data = model.matrix(mpg ~ . - 1, mtcars_test),
        select = "lambda.min")
#>                   boot1    boot2    boot3    boot4    boot5    boot6    boot7
#> Datsun 710     28.23739 22.76799 26.14163 30.79058 24.39929 26.99998 27.26240
#> Duster 360     13.82016 13.66304 16.80029 13.57099 16.12820 16.36782 17.41064
#> Merc 230       23.42902 23.03986 25.86909 26.32375 21.44036 27.21191 25.16401
#> Merc 280C      20.14410 19.69072 20.90703 17.55778 19.75631 20.14694 19.83804
#> Merc 450SLC    16.70856 16.89269 17.91436 16.95218 16.84538 16.80853 17.08464
#> Fiat 128       29.25628 24.43285 28.02294 31.80075 25.55741 28.05651 28.05296
#> Camaro Z28     13.16418 16.02415 16.37099 16.54417 15.39750 16.23831 16.48265
#> Fiat X1-9      29.90018 24.43285 28.49873 31.80075 26.49691 28.38153 28.93802
#> Lotus Europa   26.93777 21.94509 25.34544 23.40932 26.75696 27.01956 29.10974
#> Ford Pantera L 17.01820 17.81238 18.64044 20.76708 17.54120 18.20846 19.07608
#>                   boot8    boot9   boot10   boot11   boot12   boot13   boot14
#> Datsun 710     25.05114 28.32561 23.95276 26.07585 22.46841 19.18533 29.87324
#> Duster 360     18.11097 16.31502 16.11316 17.12337 10.34018 13.72910 15.09769
#> Merc 230       23.77558 31.24133 26.29606 29.53528 22.33827 25.65757 25.37163
#> Merc 280C      20.14600 20.44868 21.57800 21.48436 19.27682 25.50253 19.44269
#> Merc 450SLC    17.46243 18.37710 17.30952 18.23548 14.03156 16.25785 18.05621
#> Fiat 128       25.61771 30.91778 25.79213 28.39177 24.66792 21.02080 30.44192
#> Camaro Z28     18.19168 16.95637 16.52547 16.33836 12.05943 18.37287 12.43247
#> Fiat X1-9      26.20529 30.69151 25.90452 28.13491 25.00436 21.13954 31.59828
#> Lotus Europa   27.20930 25.25403 24.08508 23.12581 23.54140 18.38618 18.85396
#> Ford Pantera L 20.31744 19.76050 18.78042 17.44687 15.82442 16.78050 15.94418
#>                  boot15   boot16   boot17   boot18   boot19   boot20
#> Datsun 710     23.79072 18.70437 28.92896 25.02703 30.06894 25.37033
#> Duster 360     14.39689 13.70505 18.16477 15.62079 17.36808 14.42852
#> Merc 230       22.53595 20.86224 23.28615 25.46959 26.27773 21.53995
#> Merc 280C      19.45909 24.91497 19.33958 20.29652 18.98660 21.21375
#> Merc 450SLC    17.13675 16.35214 17.62744 16.52344 17.66601 16.43115
#> Fiat 128       25.21987 21.18300 29.74118 26.65879 31.10353 27.11422
#> Camaro Z28     14.69220 19.30893 17.72204 15.67979 15.85902 15.14586
#> Fiat X1-9      25.76899 21.18300 30.57515 26.95144 32.04494 27.52984
#> Lotus Europa   24.90745 21.05049 27.18999 25.75213 29.52043 24.55380
#> Ford Pantera L 16.08225 23.18600 19.91582 18.97951 18.51803 17.76267