Bootsrap-enhanced Lasso

This function implements model-consistent Lasso estimation through the bootstrap. It supports parallel processing by way of the future package, allowing the user to flexibly specify many parallelization methods. This method was developed as a variable-selection algorithm, but this package also supports making ensemble predictions on new data using the bagged Lasso models.

Usage

bolasso(
  formula,
  data,
  n.boot = 100,
  progress = TRUE,
  implement = "glmnet",
  x = NULL,
  y = NULL,
  ...
)

Arguments

formula

An optional object of class formula (or one that can be coerced to that class): a symbolic description of the model to be fitted. Can be omitted when x and y are non-missing.

data

An optional object of class data.frame that contains the modeling variables referenced in form. Can be omitted when x and y are non-missing.

n.boot

An integer specifying the number of bootstrap replicates.

progress

A boolean indicating whether to display progress across bootstrap folds.

implement

A character; either 'glmnet' or 'gamlr', specifying which Lasso implementation to utilize. For specific modeling details, see glmnet::cv.glmnet or gamlr::cv.gamlr.

x

An optional predictor matrix in lieu of form and data.

y

An optional response vector in lieu of form and data.

...

Additional parameters to pass to either glmnet::cv.glmnet or gamlr::cv.gamlr.

Value

An object of class bolasso. This object is a list of length n.boot of cv.glmnet or cv.gamlr objects.

References

Bach FR (2008). “Bolasso: model consistent Lasso estimation through the bootstrap.” CoRR, abs/0804.1302. 0804.1302, https://arxiv.org/abs/0804.1302.

See also

Examples

mtcars[, c(2, 10:11)] <- lapply(mtcars[, c(2, 10:11)], as.factor)
idx <- sample(nrow(mtcars), 22)
mtcars_train <- mtcars[idx, ]
mtcars_test <- mtcars[-idx, ]

## Formula Interface

# Train model
set.seed(123)
bolasso_form <- bolasso(
  form = mpg ~ .,
  data = mtcars_train,
  n.boot = 20,
  nfolds = 5,
  implement = "glmnet"
)
#> Loaded glmnet 4.1-4

# Extract selected variables
selected_vars(bolasso_form, threshold = 0.9, select = "lambda.min")
#> # A tibble: 2 × 2
#>   variable  mean_coef
#>   <chr>         <dbl>
#> 1 Intercept     21.9 
#> 2 wt            -3.41

# Bagged ensemble prediction on test data
predict(bolasso_form,
        new.data = mtcars_test,
        select = "lambda.min")
#>                       boot1    boot2     boot3     boot4     boot5     boot6
#> Hornet 4 Drive     24.76966 19.60619 20.439382 19.621896 19.469948 20.045485
#> Merc 240D          28.86457 23.51679 21.609953 24.538772 23.701261 23.418361
#> Merc 280C          18.67252 18.79879 17.932738 19.345382 20.253597 18.898953
#> Merc 450SE         16.58271 16.18726 15.456013 14.822995 15.208999 15.246354
#> Chrysler Imperial  10.05840 12.01500  8.988717  9.891351 10.012588  9.257491
#> Toyota Corona      34.25396 24.80906 23.327122 28.828849 25.719836 25.461839
#> AMC Javelin        22.37167 18.35202 18.320938 17.660684 16.631796 18.338503
#> Porsche 914-2      50.54142 25.75772 30.384744 30.460101 25.798958 26.843586
#> Ferrari Dino       17.88338 19.65831 22.425885 16.207442 17.596316 20.365386
#> Maserati Bora     -11.59084 14.85000 16.231427  3.331990  9.445235 13.845871
#>                       boot7     boot8     boot9   boot10    boot11   boot12
#> Hornet 4 Drive    19.796015 25.414420 20.549772 20.65564 20.865039 21.52525
#> Merc 240D         22.886404 24.785369 23.748834 25.36001 20.617868 25.22511
#> Merc 280C         19.545671 22.318143 18.221067 18.93027 19.573570 19.92463
#> Merc 450SE        15.603298 15.229572 15.668907 16.62822 15.564173 16.91157
#> Chrysler Imperial  9.502352  8.772668  7.824301 11.13482  9.375806 11.14310
#> Toyota Corona     24.537697 30.933544 26.966181 26.50520 18.824998 26.25047
#> AMC Javelin       17.434164 17.471498 19.079160 19.09039 17.224901 19.66538
#> Porsche 914-2     29.429141 25.401659 28.620241 27.54033 15.499359 27.37371
#> Ferrari Dino      22.804028 20.593344 22.099610 20.31168 15.669064 20.82321
#> Maserati Bora     15.073568 14.002159 16.609893 14.26995 16.280600 13.61503
#>                     boot13   boot14   boot15    boot16   boot17    boot18
#> Hornet 4 Drive    19.65878 19.80361 19.56249 22.568099 18.87513 20.427757
#> Merc 240D         24.33628 24.24795 22.44009 24.324560 24.36535 21.398063
#> Merc 280C         18.40324 19.59882 18.57523 20.833805 19.27964 18.141432
#> Merc 450SE        15.96904 15.07393 15.38236 15.059126 15.99985 13.946906
#> Chrysler Imperial 10.27664 10.17344 10.20471  9.658399 10.62807  4.930161
#> Toyota Corona     25.15389 24.82895 24.26704 26.323647 22.44319 25.553181
#> AMC Javelin       20.33004 17.85965 18.05082 18.487544 18.71675 18.025023
#> Porsche 914-2     26.99091 26.60371 25.46804 26.750170 24.92664 26.642284
#> Ferrari Dino      18.90008 19.11176 19.91741 21.211688 19.52710 21.405876
#> Maserati Bora     12.76373 11.17928 14.33750 13.719302 15.36366 14.988012
#>                      boot19    boot20
#> Hornet 4 Drive    20.174380 20.750624
#> Merc 240D         21.506749 23.662057
#> Merc 280C         19.145315 21.025340
#> Merc 450SE        14.528779 14.454027
#> Chrysler Imperial  8.132657  9.575352
#> Toyota Corona     24.204347 26.715676
#> AMC Javelin       17.832494 17.133417
#> Porsche 914-2     24.888632 24.889583
#> Ferrari Dino      19.974934 17.981546
#> Maserati Bora     13.544404 13.513329

## Alternal Matrix Interface

# Train model
set.seed(123)
bolasso_mat <- bolasso(
  x = model.matrix(mpg ~ . - 1, mtcars_train),
  y = mtcars_train[, 1],
  data = mtcars_train,
  n.boot = 20,
  nfolds = 5,
  implement = "glmnet"
)

# Extract selected variables
selected_vars(bolasso_mat, threshold = 0.9, select = "lambda.min")
#> # A tibble: 2 × 2
#>   variable  mean_coef
#>   <chr>         <dbl>
#> 1 Intercept     21.9 
#> 2 wt            -3.41

# Bagged ensemble prediction on test data
predict(bolasso_mat,
        new.data = model.matrix(mpg ~ . - 1, mtcars_test),
        select = "lambda.min")
#>                       boot1    boot2     boot3     boot4     boot5     boot6
#> Hornet 4 Drive     24.76966 19.60619 20.439382 19.621896 19.469948 20.045485
#> Merc 240D          28.86457 23.51679 21.609953 24.538772 23.701261 23.418361
#> Merc 280C          18.67252 18.79879 17.932738 19.345382 20.253597 18.898953
#> Merc 450SE         16.58271 16.18726 15.456013 14.822995 15.208999 15.246354
#> Chrysler Imperial  10.05840 12.01500  8.988717  9.891351 10.012588  9.257491
#> Toyota Corona      34.25396 24.80906 23.327122 28.828849 25.719836 25.461839
#> AMC Javelin        22.37167 18.35202 18.320938 17.660684 16.631796 18.338503
#> Porsche 914-2      50.54142 25.75772 30.384744 30.460101 25.798958 26.843586
#> Ferrari Dino       17.88338 19.65831 22.425885 16.207442 17.596316 20.365386
#> Maserati Bora     -11.59084 14.85000 16.231427  3.331990  9.445235 13.845871
#>                       boot7     boot8     boot9   boot10    boot11   boot12
#> Hornet 4 Drive    19.796015 25.414420 20.549772 20.65564 20.865039 21.52525
#> Merc 240D         22.886404 24.785369 23.748834 25.36001 20.617868 25.22511
#> Merc 280C         19.545671 22.318143 18.221067 18.93027 19.573570 19.92463
#> Merc 450SE        15.603298 15.229572 15.668907 16.62822 15.564173 16.91157
#> Chrysler Imperial  9.502352  8.772668  7.824301 11.13482  9.375806 11.14310
#> Toyota Corona     24.537697 30.933544 26.966181 26.50520 18.824998 26.25047
#> AMC Javelin       17.434164 17.471498 19.079160 19.09039 17.224901 19.66538
#> Porsche 914-2     29.429141 25.401659 28.620241 27.54033 15.499359 27.37371
#> Ferrari Dino      22.804028 20.593344 22.099610 20.31168 15.669064 20.82321
#> Maserati Bora     15.073568 14.002159 16.609893 14.26995 16.280600 13.61503
#>                     boot13   boot14   boot15    boot16   boot17    boot18
#> Hornet 4 Drive    19.65878 19.80361 19.56249 22.568099 18.87513 20.427757
#> Merc 240D         24.33628 24.24795 22.44009 24.324560 24.36535 21.398063
#> Merc 280C         18.40324 19.59882 18.57523 20.833805 19.27964 18.141432
#> Merc 450SE        15.96904 15.07393 15.38236 15.059126 15.99985 13.946906
#> Chrysler Imperial 10.27664 10.17344 10.20471  9.658399 10.62807  4.930161
#> Toyota Corona     25.15389 24.82895 24.26704 26.323647 22.44319 25.553181
#> AMC Javelin       20.33004 17.85965 18.05082 18.487544 18.71675 18.025023
#> Porsche 914-2     26.99091 26.60371 25.46804 26.750170 24.92664 26.642284
#> Ferrari Dino      18.90008 19.11176 19.91741 21.211688 19.52710 21.405876
#> Maserati Bora     12.76373 11.17928 14.33750 13.719302 15.36366 14.988012
#>                      boot19    boot20
#> Hornet 4 Drive    20.174380 20.750624
#> Merc 240D         21.506749 23.662057
#> Merc 280C         19.145315 21.025340
#> Merc 450SE        14.528779 14.454027
#> Chrysler Imperial  8.132657  9.575352
#> Toyota Corona     24.204347 26.715676
#> AMC Javelin       17.832494 17.133417
#> Porsche 914-2     24.888632 24.889583
#> Ferrari Dino      19.974934 17.981546
#> Maserati Bora     13.544404 13.513329