Bootsrap-enhanced Lasso

This function implements model-consistent Lasso estimation through the bootstrap. It supports parallel processing by way of the future package, allowing the user to flexibly specify many parallelization methods. This method was developed as a variable-selection algorithm, but this package also supports making ensemble predictions on new data using the bagged Lasso models.

Usage

bolasso(
  form,
  data,
  n.boot = 100,
  progress = TRUE,
  implement = "glmnet",
  x = NULL,
  y = NULL,
  ...
)

Arguments

form

An optional object of class formula (or one that can be coerced to that class): a symbolic description of the model to be fitted. Can be omitted when x and y are non-missing.

data

An optional object of class data.frame that contains the modeling variables referenced in form. Can be omitted when x and y are non-missing.

n.boot

An integer specifying the number of bootstrap replicates.

progress

A boolean indicating whether to display progress across bootstrap folds.

implement

A character; either 'glmnet' or 'gamlr', specifying which Lasso implementation to utilize. For specific modeling details, see glmnet::cv.glmnet or gamlr::cv.gamlr.

x

An optional predictor matrix in lieu of form and data.

y

An optional response vector in lieu of form and data.

...

Additional parameters to pass to either glmnet::cv.glmnet or gamlr::cv.gamlr.

Value

An object of class bolasso. This object is a list of length n.boot of cv.glmnet or cv.gamlr objects.

References

Bach FR (2008). “Bolasso: model consistent Lasso estimation through the bootstrap.” CoRR, abs/0804.1302. 0804.1302, http://arxiv.org/abs/0804.1302.

See also

Examples

mtcars[, c(2, 10:11)] <- lapply(mtcars[, c(2, 10:11)], as.factor)
idx <- sample(nrow(mtcars), 22)
mtcars_train <- mtcars[idx, ]
mtcars_test <- mtcars[-idx, ]

## Formula Interface

# Train model
set.seed(123)
bolasso_form <- bolasso(
  form = mpg ~ .,
  data = mtcars_train,
  n.boot = 20,
  nfolds = 5,
  implement = "glmnet"
)
#> Loaded glmnet 4.1-3

# Extract selected variables
selected_vars(bolasso_form, threshold = 0.9, select = "lambda.min")
#> # A tibble: 1 × 2
#>   variable  mean_coef
#>   <chr>         <dbl>
#> 1 Intercept      21.3

# Bagged ensemble prediction on test data
predict(bolasso_form,
        new.data = mtcars_test,
        select = "lambda.min")
#>                         boot1    boot2     boot3    boot4     boot5    boot6
#> Mazda RX4           23.634490 19.46660 23.296169 20.56045 21.677787 18.51405
#> Mazda RX4 Wag       21.759530 19.26401 22.570195 20.28997 20.796195 18.11381
#> Merc 230            19.949583 25.00192 22.993488 21.85809 22.031604 22.60310
#> Merc 450SE          14.061720 17.32012 14.736112 16.24978 14.891545 15.15425
#> Cadillac Fleetwood   3.657769 14.69768 10.404302 13.56352  9.415715 12.91504
#> Lincoln Continental  2.083518 14.66249  9.560039 13.36735  8.701481 12.48709
#> Honda Civic         32.261736 26.78052 31.015617 25.71117 28.397999 25.67822
#> Toyota Corolla      33.378988 26.64524 28.374035 25.39437 27.436726 25.13162
#> Pontiac Firebird    12.480466 16.43225 17.112815 15.75973 15.143321 15.58482
#> Maserati Bora       11.670668 17.50094 12.903572 16.73753 13.832950 16.54293
#>                         boot7    boot8    boot9    boot10   boot11   boot12
#> Mazda RX4           34.859714 20.76775 13.94173 20.099603 20.82233 21.03660
#> Mazda RX4 Wag       31.828643 22.97999 13.94173 18.564172 21.39068 21.03660
#> Merc 230            13.558236 39.90981 22.67618 24.778018 29.94412 23.42378
#> Merc 450SE          12.278013 14.58890 15.88135 15.212310 16.65399 17.12126
#> Cadillac Fleetwood   1.118233 12.46900 13.64593  6.650690 14.96588 15.00030
#> Lincoln Continental -2.691499 12.75933 13.68024  5.405353 14.80349 14.90787
#> Honda Civic         45.748121 42.91090 23.72618 32.063378 31.97065 27.45300
#> Toyota Corolla      37.188355 35.40062 23.65791 27.904968 30.90393 25.73329
#> Pontiac Firebird    24.697402 16.36203 14.65126 16.037816 17.99115 17.25733
#> Maserati Bora       -2.499668 16.08049 15.96478 12.181541 19.50916 15.47433
#>                       boot13   boot14   boot15   boot16   boot17    boot18
#> Mazda RX4           19.25972 23.42131 21.03323 21.65285 21.09482 23.370558
#> Mazda RX4 Wag       18.93583 23.34711 20.58433 21.01387 20.60457 22.272929
#> Merc 230            23.96029 22.55895 21.45382 20.81391 23.04185 20.529456
#> Merc 450SE          14.59418 17.46229 18.04171 15.63514 16.81188 14.603523
#> Cadillac Fleetwood  13.07610 15.69865 15.80769 11.83318 14.00928  7.451267
#> Lincoln Continental 12.85509 15.07990 15.43867 11.05912 13.46115  6.606850
#> Honda Civic         25.90994 28.43412 24.42565 26.13187 26.91145 29.105203
#> Toyota Corolla      23.60458 27.63156 23.95685 25.14113 26.21080 27.973030
#> Pontiac Firebird    16.90589 17.81182 18.46915 16.36797 17.35125 14.385401
#> Maserati Bora       16.31376 11.78768 17.94995 12.58402 14.46218 14.434864
#>                        boot19    boot20
#> Mazda RX4           23.326044 19.963582
#> Mazda RX4 Wag       21.932690 19.463530
#> Merc 230            21.105021 27.853275
#> Merc 450SE          13.781404 14.542168
#> Cadillac Fleetwood   5.599487  9.563645
#> Lincoln Continental  4.475102  8.551324
#> Honda Civic         30.542381 29.614284
#> Toyota Corolla      29.114558 31.300915
#> Pontiac Firebird    13.797473 14.154015
#> Maserati Bora       13.200197 14.713361

## Alternal Matrix Interface

# Train model
set.seed(123)
bolasso_mat <- bolasso(
  x = model.matrix(mpg ~ . - 1, mtcars_train),
  y = mtcars_train[, 1],
  data = mtcars_train,
  n.boot = 20,
  nfolds = 5,
  implement = "glmnet"
)

# Extract selected variables
selected_vars(bolasso_mat, threshold = 0.9, select = "lambda.min")
#> # A tibble: 1 × 2
#>   variable  mean_coef
#>   <chr>         <dbl>
#> 1 Intercept      21.3

# Bagged ensemble prediction on test data
predict(bolasso_mat,
        new.data = model.matrix(mpg ~ . - 1, mtcars_test),
        select = "lambda.min")
#>                         boot1    boot2     boot3    boot4     boot5    boot6
#> Mazda RX4           23.634490 19.46660 23.296169 20.56045 21.677787 18.51405
#> Mazda RX4 Wag       21.759530 19.26401 22.570195 20.28997 20.796195 18.11381
#> Merc 230            19.949583 25.00192 22.993488 21.85809 22.031604 22.60310
#> Merc 450SE          14.061720 17.32012 14.736112 16.24978 14.891545 15.15425
#> Cadillac Fleetwood   3.657769 14.69768 10.404302 13.56352  9.415715 12.91504
#> Lincoln Continental  2.083518 14.66249  9.560039 13.36735  8.701481 12.48709
#> Honda Civic         32.261736 26.78052 31.015617 25.71117 28.397999 25.67822
#> Toyota Corolla      33.378988 26.64524 28.374035 25.39437 27.436726 25.13162
#> Pontiac Firebird    12.480466 16.43225 17.112815 15.75973 15.143321 15.58482
#> Maserati Bora       11.670668 17.50094 12.903572 16.73753 13.832950 16.54293
#>                         boot7    boot8    boot9    boot10   boot11   boot12
#> Mazda RX4           34.859714 20.76775 13.94173 20.099603 20.82233 21.03660
#> Mazda RX4 Wag       31.828643 22.97999 13.94173 18.564172 21.39068 21.03660
#> Merc 230            13.558236 39.90981 22.67618 24.778018 29.94412 23.42378
#> Merc 450SE          12.278013 14.58890 15.88135 15.212310 16.65399 17.12126
#> Cadillac Fleetwood   1.118233 12.46900 13.64593  6.650690 14.96588 15.00030
#> Lincoln Continental -2.691499 12.75933 13.68024  5.405353 14.80349 14.90787
#> Honda Civic         45.748121 42.91090 23.72618 32.063378 31.97065 27.45300
#> Toyota Corolla      37.188355 35.40062 23.65791 27.904968 30.90393 25.73329
#> Pontiac Firebird    24.697402 16.36203 14.65126 16.037816 17.99115 17.25733
#> Maserati Bora       -2.499668 16.08049 15.96478 12.181541 19.50916 15.47433
#>                       boot13   boot14   boot15   boot16   boot17    boot18
#> Mazda RX4           19.25972 23.42131 21.03323 21.65285 21.09482 23.370558
#> Mazda RX4 Wag       18.93583 23.34711 20.58433 21.01387 20.60457 22.272929
#> Merc 230            23.96029 22.55895 21.45382 20.81391 23.04185 20.529456
#> Merc 450SE          14.59418 17.46229 18.04171 15.63514 16.81188 14.603523
#> Cadillac Fleetwood  13.07610 15.69865 15.80769 11.83318 14.00928  7.451267
#> Lincoln Continental 12.85509 15.07990 15.43867 11.05912 13.46115  6.606850
#> Honda Civic         25.90994 28.43412 24.42565 26.13187 26.91145 29.105203
#> Toyota Corolla      23.60458 27.63156 23.95685 25.14113 26.21080 27.973030
#> Pontiac Firebird    16.90589 17.81182 18.46915 16.36797 17.35125 14.385401
#> Maserati Bora       16.31376 11.78768 17.94995 12.58402 14.46218 14.434864
#>                        boot19    boot20
#> Mazda RX4           23.326044 19.963582
#> Mazda RX4 Wag       21.932690 19.463530
#> Merc 230            21.105021 27.853275
#> Merc 450SE          13.781404 14.542168
#> Cadillac Fleetwood   5.599487  9.563645
#> Lincoln Continental  4.475102  8.551324
#> Honda Civic         30.542381 29.614284
#> Toyota Corolla      29.114558 31.300915
#> Pontiac Firebird    13.797473 14.154015
#> Maserati Bora       13.200197 14.713361