Bootsrap-enhanced Lasso

This function implements model-consistent Lasso estimation through the bootstrap. It supports parallel processing by way of the future package, allowing the user to flexibly specify many parallelization methods. This method was developed as a variable-selection algorithm, but this package also supports making ensemble predictions on new data using the bagged Lasso models.

Usage

bolasso(
  form,
  data,
  n.boot = 100,
  progress = TRUE,
  implement = "glmnet",
  x = NULL,
  y = NULL,
  ...
)

Arguments

form

An optional object of class formula (or one that can be coerced to that class): a symbolic description of the model to be fitted. Can be omitted when x and y are non-missing.

data

An optional object of class data.frame that contains the modeling variables referenced in form. Can be omitted when x and y are non-missing.

n.boot

An integer specifying the number of bootstrap replicates.

progress

A boolean indicating whether to display progress across bootstrap folds.

implement

A character; either 'glmnet' or 'gamlr', specifying which Lasso implementation to utilize. For specific modeling details, see glmnet::cv.glmnet or gamlr::cv.gamlr.

x

An optional predictor matrix in lieu of form and data.

y

An optional response vector in lieu of form and data.

...

Additional parameters to pass to either glmnet::cv.glmnet or gamlr::cv.gamlr.

Value

An object of class bolasso. This object is a list of length n.boot of cv.glmnet or cv.gamlr objects.

References

Bach FR (2008). “Bolasso: model consistent Lasso estimation through the bootstrap.” CoRR, abs/0804.1302. 0804.1302, https://arxiv.org/abs/0804.1302.

See also

Examples

mtcars[, c(2, 10:11)] <- lapply(mtcars[, c(2, 10:11)], as.factor)
idx <- sample(nrow(mtcars), 22)
mtcars_train <- mtcars[idx, ]
mtcars_test <- mtcars[-idx, ]

## Formula Interface

# Train model
set.seed(123)
bolasso_form <- bolasso(
  form = mpg ~ .,
  data = mtcars_train,
  n.boot = 20,
  nfolds = 5,
  implement = "glmnet"
)
#> Loaded glmnet 4.1-3

# Extract selected variables
selected_vars(bolasso_form, threshold = 0.9, select = "lambda.min")
#> # A tibble: 2 × 2
#>   variable  mean_coef
#>   <chr>         <dbl>
#> 1 Intercept     26.2 
#> 2 wt            -2.31

# Bagged ensemble prediction on test data
predict(bolasso_form,
        new.data = mtcars_test,
        select = "lambda.min")
#>                        boot1    boot2    boot3     boot4    boot5    boot6
#> Mazda RX4 Wag       21.76404 21.18122 20.83347 23.641015 21.24783 22.17019
#> Hornet Sportabout   17.48812 19.78114 18.53847 15.576519 21.67950 19.00244
#> Merc 230            30.73895 26.01162 22.99075 16.859413 23.84509 24.02592
#> Cadillac Fleetwood  16.40633 14.73581 13.97884  3.488837 13.43936 14.93018
#> Lincoln Continental 15.44837 14.33812 13.42902  3.438867 12.81341 14.59516
#> Dodge Challenger    16.42510 20.37255 18.74941 15.351451 21.91622 19.41979
#> AMC Javelin         19.53651 20.53476 18.94146 14.244904 22.13543 19.49011
#> Camaro Z28          14.89375 13.61969 16.53794 11.520334 16.36679 15.30743
#> Fiat X1-9           29.14683 26.39164 26.19034 34.328702 24.45176 27.02529
#> Ferrari Dino        17.64823 19.05746 20.05230 31.718382 20.36674 20.99762
#>                        boot7    boot8    boot9    boot10    boot11    boot12
#> Mazda RX4 Wag       19.78346 21.55597 22.88456 22.038341 22.165480 21.195188
#> Hornet Sportabout   17.26624 19.42774 17.09754 19.265093 17.509247 19.426048
#> Merc 230            24.41449 23.40554 29.75001 22.269295 22.183296 21.501687
#> Cadillac Fleetwood  14.25118 12.55599 15.37018  8.852272  8.942520  8.699510
#> Lincoln Continental 14.00112 11.88878 15.05006  7.846180  8.129574  7.759599
#> Dodge Challenger    17.68346 19.70814 15.92821 19.316448 17.472153 19.231834
#> AMC Javelin         17.85545 19.93717 17.61460 19.725125 17.813714 19.651716
#> Camaro Z28          16.06079 15.56162 13.75048 14.878623 14.153410 15.343061
#> Fiat X1-9           26.46848 25.85492 27.63789 28.554612 27.395482 26.192294
#> Ferrari Dino        19.17849 19.83334 19.62443 22.457227 20.201550 21.191347
#>                       boot13   boot14   boot15   boot16   boot17    boot18
#> Mazda RX4 Wag       22.27020 19.27667 21.34405 22.06339 21.82334 20.011358
#> Hornet Sportabout   16.17989 19.55355 18.65733 19.78015 18.66327 20.384878
#> Merc 230            23.87149 22.19578 23.76472 21.98604 22.16649 22.521305
#> Cadillac Fleetwood  14.71869 12.47680 12.76120 13.14195 15.72726  8.945469
#> Lincoln Continental 14.71869 11.90752 12.03181 12.48434 15.23285  8.521675
#> Dodge Challenger    16.17989 19.98405 18.71226 19.55873 19.68803 19.833105
#> AMC Javelin         16.17989 20.14789 19.09031 19.86662 19.72290 20.299125
#> Camaro Z28          14.71869 14.25915 16.01434 18.13988 14.61359 14.356170
#> Fiat X1-9           30.77723 23.74329 26.99299 26.52527 24.99467 24.202800
#> Ferrari Dino        23.08562 20.18886 20.39036 22.26600 21.32290 21.154947
#>                       boot19   boot20
#> Mazda RX4 Wag       20.30447 21.45415
#> Hornet Sportabout   18.46757 19.28401
#> Merc 230            25.57238 23.28621
#> Cadillac Fleetwood  15.46526 14.88158
#> Lincoln Continental 15.06313 14.33929
#> Dodge Challenger    18.75481 19.62629
#> AMC Javelin         18.87333 19.79356
#> Camaro Z28          16.79323 16.85673
#> Fiat X1-9           27.74135 26.01588
#> Ferrari Dino        19.41403 20.36153

## Alternal Matrix Interface

# Train model
set.seed(123)
bolasso_mat <- bolasso(
  x = model.matrix(mpg ~ . - 1, mtcars_train),
  y = mtcars_train[, 1],
  data = mtcars_train,
  n.boot = 20,
  nfolds = 5,
  implement = "glmnet"
)

# Extract selected variables
selected_vars(bolasso_mat, threshold = 0.9, select = "lambda.min")
#> # A tibble: 2 × 2
#>   variable  mean_coef
#>   <chr>         <dbl>
#> 1 Intercept     26.2 
#> 2 wt            -2.31

# Bagged ensemble prediction on test data
predict(bolasso_mat,
        new.data = model.matrix(mpg ~ . - 1, mtcars_test),
        select = "lambda.min")
#>                        boot1    boot2    boot3     boot4    boot5    boot6
#> Mazda RX4 Wag       21.76404 21.18122 20.83347 23.641015 21.24783 22.17019
#> Hornet Sportabout   17.48812 19.78114 18.53847 15.576519 21.67950 19.00244
#> Merc 230            30.73895 26.01162 22.99075 16.859413 23.84509 24.02592
#> Cadillac Fleetwood  16.40633 14.73581 13.97884  3.488837 13.43936 14.93018
#> Lincoln Continental 15.44837 14.33812 13.42902  3.438867 12.81341 14.59516
#> Dodge Challenger    16.42510 20.37255 18.74941 15.351451 21.91622 19.41979
#> AMC Javelin         19.53651 20.53476 18.94146 14.244904 22.13543 19.49011
#> Camaro Z28          14.89375 13.61969 16.53794 11.520334 16.36679 15.30743
#> Fiat X1-9           29.14683 26.39164 26.19034 34.328702 24.45176 27.02529
#> Ferrari Dino        17.64823 19.05746 20.05230 31.718382 20.36674 20.99762
#>                        boot7    boot8    boot9    boot10    boot11    boot12
#> Mazda RX4 Wag       19.78346 21.55597 22.88456 22.038341 22.165480 21.195188
#> Hornet Sportabout   17.26624 19.42774 17.09754 19.265093 17.509247 19.426048
#> Merc 230            24.41449 23.40554 29.75001 22.269295 22.183296 21.501687
#> Cadillac Fleetwood  14.25118 12.55599 15.37018  8.852272  8.942520  8.699510
#> Lincoln Continental 14.00112 11.88878 15.05006  7.846180  8.129574  7.759599
#> Dodge Challenger    17.68346 19.70814 15.92821 19.316448 17.472153 19.231834
#> AMC Javelin         17.85545 19.93717 17.61460 19.725125 17.813714 19.651716
#> Camaro Z28          16.06079 15.56162 13.75048 14.878623 14.153410 15.343061
#> Fiat X1-9           26.46848 25.85492 27.63789 28.554612 27.395482 26.192294
#> Ferrari Dino        19.17849 19.83334 19.62443 22.457227 20.201550 21.191347
#>                       boot13   boot14   boot15   boot16   boot17    boot18
#> Mazda RX4 Wag       22.27020 19.27667 21.34405 22.06339 21.82334 20.011358
#> Hornet Sportabout   16.17989 19.55355 18.65733 19.78015 18.66327 20.384878
#> Merc 230            23.87149 22.19578 23.76472 21.98604 22.16649 22.521305
#> Cadillac Fleetwood  14.71869 12.47680 12.76120 13.14195 15.72726  8.945469
#> Lincoln Continental 14.71869 11.90752 12.03181 12.48434 15.23285  8.521675
#> Dodge Challenger    16.17989 19.98405 18.71226 19.55873 19.68803 19.833105
#> AMC Javelin         16.17989 20.14789 19.09031 19.86662 19.72290 20.299125
#> Camaro Z28          14.71869 14.25915 16.01434 18.13988 14.61359 14.356170
#> Fiat X1-9           30.77723 23.74329 26.99299 26.52527 24.99467 24.202800
#> Ferrari Dino        23.08562 20.18886 20.39036 22.26600 21.32290 21.154947
#>                       boot19   boot20
#> Mazda RX4 Wag       20.30447 21.45415
#> Hornet Sportabout   18.46757 19.28401
#> Merc 230            25.57238 23.28621
#> Cadillac Fleetwood  15.46526 14.88158
#> Lincoln Continental 15.06313 14.33929
#> Dodge Challenger    18.75481 19.62629
#> AMC Javelin         18.87333 19.79356
#> Camaro Z28          16.79323 16.85673
#> Fiat X1-9           27.74135 26.01588
#> Ferrari Dino        19.41403 20.36153