Bootsrap-enhanced Lasso

This function implements model-consistent Lasso estimation through the bootstrap. It supports parallel processing by way of the future package, allowing the user to flexibly specify many parallelization methods. This method was developed as a variable-selection algorithm, but this package also supports making ensemble predictions on new data using the bagged Lasso models.

Usage

bolasso(
  form,
  data,
  n.boot = 100,
  progress = TRUE,
  implement = "glmnet",
  x = NULL,
  y = NULL,
  ...
)

Arguments

form

An optional object of class formula (or one that can be coerced to that class): a symbolic description of the model to be fitted. Can be omitted when x and y are non-missing.

data

An optional object of class data.frame that contains the modeling variables referenced in form. Can be omitted when x and y are non-missing.

n.boot

An integer specifying the number of bootstrap replicates.

progress

A boolean indicating whether to display progress across bootstrap folds.

implement

A character; either 'glmnet' or 'gamlr', specifying which Lasso implementation to utilize. For specific modeling details, see glmnet::cv.glmnet or gamlr::cv.gamlr.

x

An optional predictor matrix in lieu of form and data.

y

An optional response vector in lieu of form and data.

...

Additional parameters to pass to either glmnet::cv.glmnet or gamlr::cv.gamlr.

Value

An object of class bolasso. This object is a list of length n.boot of cv.glmnet or cv.gamlr objects.

References

Bach FR (2008). “Bolasso: model consistent Lasso estimation through the bootstrap.” CoRR, abs/0804.1302. 0804.1302, https://arxiv.org/abs/0804.1302.

See also

Examples

mtcars[, c(2, 10:11)] <- lapply(mtcars[, c(2, 10:11)], as.factor)
idx <- sample(nrow(mtcars), 22)
mtcars_train <- mtcars[idx, ]
mtcars_test <- mtcars[-idx, ]

## Formula Interface

# Train model
set.seed(123)
bolasso_form <- bolasso(
  form = mpg ~ .,
  data = mtcars_train,
  n.boot = 20,
  nfolds = 5,
  implement = "glmnet"
)
#> Loaded glmnet 4.1-3

# Extract selected variables
selected_vars(bolasso_form, threshold = 0.9, select = "lambda.min")
#> # A tibble: 1 × 2
#>   variable  mean_coef
#>   <chr>         <dbl>
#> 1 Intercept      26.3

# Bagged ensemble prediction on test data
predict(bolasso_form,
        new.data = mtcars_test,
        select = "lambda.min")
#>                      boot1    boot2    boot3    boot4    boot5    boot6
#> Mazda RX4 Wag     19.99757 19.98025 23.12462 20.41575 20.73373 21.84665
#> Datsun 710        24.28814 26.45366 26.65587 31.91484 25.68585 30.32397
#> Duster 360        13.36547 16.39809 15.97934 13.85490 17.03136 14.03091
#> Merc 450SE        14.86574 15.54451 15.37442 14.67944 15.80518 15.41195
#> Chrysler Imperial 11.10688 11.07426 11.00613 10.07774 11.71491 10.81746
#> Fiat 128          25.28465 27.11139 27.35899 32.95893 26.17038 30.54121
#> Fiat X1-9         25.66794 27.93003 28.13080 33.66601 26.97411 31.02097
#> Porsche 914-2     25.53249 27.03097 26.03048 25.99840 26.24023 26.34839
#> Ford Pantera L    16.74306 17.43178 18.74765 17.73403 18.15968 18.88833
#> Volvo 142E        25.88578 24.86254 25.10657 27.37477 24.21925 29.49119
#>                      boot7    boot8    boot9    boot10   boot11   boot12
#> Mazda RX4 Wag     27.09102 20.73109 20.66918 16.136986 19.69954 21.58850
#> Datsun 710        37.11596 25.74299 22.88049 21.655971 28.96567 31.05141
#> Duster 360        14.42402 16.86136 13.21075  9.649251 15.00978 11.25844
#> Merc 450SE        15.91199 15.60627 14.67814 17.338625 16.76451 17.65888
#> Chrysler Imperial 10.08411 10.37117 10.46376 11.388257 10.25530 11.34653
#> Fiat 128          34.19871 26.44595 24.28139 23.444616 29.94375 31.80935
#> Fiat X1-9         36.10130 27.43603 24.76877 23.594503 30.64241 31.80943
#> Porsche 914-2     29.27946 26.43436 27.94222 25.282004 29.24082 30.07835
#> Ford Pantera L    19.67139 18.17663 16.26213 15.596082 15.64625 22.55192
#> Volvo 142E        34.24428 23.87348 25.25135 23.352621 27.30865 31.92061
#>                     boot13   boot14   boot15   boot16   boot17   boot18
#> Mazda RX4 Wag     21.28049 21.07309 21.18232 26.86065 20.08305 25.80368
#> Datsun 710        27.58922 26.80002 25.43289 33.39117 24.93316 29.05321
#> Duster 360        15.10295 16.99624 17.55462 11.47752 15.63087 13.02393
#> Merc 450SE        14.36921 15.91549 16.93543 17.20479 16.00669 16.08140
#> Chrysler Imperial 10.01412 10.69465 12.70735 11.41107  9.89071 10.38507
#> Fiat 128          28.25579 27.54610 26.10457 33.52482 25.75964 30.22219
#> Fiat X1-9         29.04131 28.51249 26.85574 33.79090 26.65166 30.69535
#> Porsche 914-2     26.96579 26.10995 25.96767 32.81539 27.59730 27.62764
#> Ford Pantera L    18.34748 18.23786 18.45518 23.21856 16.67998 18.01385
#> Volvo 142E        26.04145 24.93972 23.93250 33.70703 25.41200 27.66613
#>                     boot19   boot20
#> Mazda RX4 Wag     22.87049 15.44686
#> Datsun 710        29.76078 20.40571
#> Duster 360        14.71074 12.46453
#> Merc 450SE        14.32996 17.04984
#> Chrysler Imperial 10.39206 11.01615
#> Fiat 128          30.77747 18.33438
#> Fiat X1-9         31.17178 19.30155
#> Porsche 914-2     26.73666 25.24888
#> Ford Pantera L    17.91646 16.53628
#> Volvo 142E        28.32714 27.72665

## Alternal Matrix Interface

# Train model
set.seed(123)
bolasso_mat <- bolasso(
  x = model.matrix(mpg ~ . - 1, mtcars_train),
  y = mtcars_train[, 1],
  data = mtcars_train,
  n.boot = 20,
  nfolds = 5,
  implement = "glmnet"
)

# Extract selected variables
selected_vars(bolasso_mat, threshold = 0.9, select = "lambda.min")
#> # A tibble: 1 × 2
#>   variable  mean_coef
#>   <chr>         <dbl>
#> 1 Intercept      26.3

# Bagged ensemble prediction on test data
predict(bolasso_mat,
        new.data = model.matrix(mpg ~ . - 1, mtcars_test),
        select = "lambda.min")
#>                      boot1    boot2    boot3    boot4    boot5    boot6
#> Mazda RX4 Wag     19.99757 19.98025 23.12462 20.41575 20.73373 21.84665
#> Datsun 710        24.28814 26.45366 26.65587 31.91484 25.68585 30.32397
#> Duster 360        13.36547 16.39809 15.97934 13.85490 17.03136 14.03091
#> Merc 450SE        14.86574 15.54451 15.37442 14.67944 15.80518 15.41195
#> Chrysler Imperial 11.10688 11.07426 11.00613 10.07774 11.71491 10.81746
#> Fiat 128          25.28465 27.11139 27.35899 32.95893 26.17038 30.54121
#> Fiat X1-9         25.66794 27.93003 28.13080 33.66601 26.97411 31.02097
#> Porsche 914-2     25.53249 27.03097 26.03048 25.99840 26.24023 26.34839
#> Ford Pantera L    16.74306 17.43178 18.74765 17.73403 18.15968 18.88833
#> Volvo 142E        25.88578 24.86254 25.10657 27.37477 24.21925 29.49119
#>                      boot7    boot8    boot9    boot10   boot11   boot12
#> Mazda RX4 Wag     27.09102 20.73109 20.66918 16.136986 19.69954 21.58850
#> Datsun 710        37.11596 25.74299 22.88049 21.655971 28.96567 31.05141
#> Duster 360        14.42402 16.86136 13.21075  9.649251 15.00978 11.25844
#> Merc 450SE        15.91199 15.60627 14.67814 17.338625 16.76451 17.65888
#> Chrysler Imperial 10.08411 10.37117 10.46376 11.388257 10.25530 11.34653
#> Fiat 128          34.19871 26.44595 24.28139 23.444616 29.94375 31.80935
#> Fiat X1-9         36.10130 27.43603 24.76877 23.594503 30.64241 31.80943
#> Porsche 914-2     29.27946 26.43436 27.94222 25.282004 29.24082 30.07835
#> Ford Pantera L    19.67139 18.17663 16.26213 15.596082 15.64625 22.55192
#> Volvo 142E        34.24428 23.87348 25.25135 23.352621 27.30865 31.92061
#>                     boot13   boot14   boot15   boot16   boot17   boot18
#> Mazda RX4 Wag     21.28049 21.07309 21.18232 26.86065 20.08305 25.80368
#> Datsun 710        27.58922 26.80002 25.43289 33.39117 24.93316 29.05321
#> Duster 360        15.10295 16.99624 17.55462 11.47752 15.63087 13.02393
#> Merc 450SE        14.36921 15.91549 16.93543 17.20479 16.00669 16.08140
#> Chrysler Imperial 10.01412 10.69465 12.70735 11.41107  9.89071 10.38507
#> Fiat 128          28.25579 27.54610 26.10457 33.52482 25.75964 30.22219
#> Fiat X1-9         29.04131 28.51249 26.85574 33.79090 26.65166 30.69535
#> Porsche 914-2     26.96579 26.10995 25.96767 32.81539 27.59730 27.62764
#> Ford Pantera L    18.34748 18.23786 18.45518 23.21856 16.67998 18.01385
#> Volvo 142E        26.04145 24.93972 23.93250 33.70703 25.41200 27.66613
#>                     boot19   boot20
#> Mazda RX4 Wag     22.87049 15.44686
#> Datsun 710        29.76078 20.40571
#> Duster 360        14.71074 12.46453
#> Merc 450SE        14.32996 17.04984
#> Chrysler Imperial 10.39206 11.01615
#> Fiat 128          30.77747 18.33438
#> Fiat X1-9         31.17178 19.30155
#> Porsche 914-2     26.73666 25.24888
#> Ford Pantera L    17.91646 16.53628
#> Volvo 142E        28.32714 27.72665