Bootsrap-enhanced Lasso

This function implements model-consistent Lasso estimation through the bootstrap. It supports parallel processing by way of the future package, allowing the user to flexibly specify many parallelization methods. This method was developed as a variable-selection algorithm, but this package also supports making ensemble predictions on new data using the bagged Lasso models.

Usage

bolasso(
  form,
  data,
  n.boot = 100,
  progress = TRUE,
  implement = "glmnet",
  x = NULL,
  y = NULL,
  ...
)

Arguments

form

An optional object of class formula (or one that can be coerced to that class): a symbolic description of the model to be fitted. Can be omitted when x and y are non-missing.

data

An optional object of class data.frame that contains the modeling variables referenced in form. Can be omitted when x and y are non-missing.

n.boot

An integer specifying the number of bootstrap replicates.

progress

A boolean indicating whether to display progress across bootstrap folds.

implement

A character; either 'glmnet' or 'gamlr', specifying which Lasso implementation to utilize. For specific modeling details, see glmnet::cv.glmnet or gamlr::cv.gamlr.

x

An optional predictor matrix in lieu of form and data.

y

An optional response vector in lieu of form and data.

...

Additional parameters to pass to either glmnet::cv.glmnet or gamlr::cv.gamlr.

Value

An object of class bolasso. This object is a list of length n.boot of cv.glmnet or cv.gamlr objects.

References

Bach FR (2008). “Bolasso: model consistent Lasso estimation through the bootstrap.” CoRR, abs/0804.1302. 0804.1302, http://arxiv.org/abs/0804.1302.

See also

Examples

mtcars[, c(2, 10:11)] <- lapply(mtcars[, c(2, 10:11)], as.factor)
idx <- sample(nrow(mtcars), 22)
mtcars_train <- mtcars[idx, ]
mtcars_test <- mtcars[-idx, ]

## Formula Interface

# Train model
set.seed(123)
bolasso_form <- bolasso(
  form = mpg ~ .,
  data = mtcars_train,
  n.boot = 20,
  nfolds = 5,
  implement = "glmnet"
)
#> Loaded glmnet 4.1-3

# Extract selected variables
selected_vars(bolasso_form, threshold = 0.9, select = "lambda.min")
#> # A tibble: 2 × 2
#>   variable  mean_coef
#>   <chr>         <dbl>
#> 1 Intercept     23.1 
#> 2 wt            -3.68

# Bagged ensemble prediction on test data
predict(bolasso_form,
        new.data = mtcars_test,
        select = "lambda.min")
#>                         boot1    boot2    boot3     boot4     boot5    boot6
#> Mazda RX4           21.202865 24.06137 20.37871 19.693430 22.115000 22.88248
#> Valiant             20.825762 20.27784 14.05869 21.108939 20.806772 19.94063
#> Duster 360          15.241608 12.73562 11.37251 16.573997 16.048799 16.93248
#> Merc 230            27.293281 23.38017 26.61376 27.069320 25.023952 21.55425
#> Merc 450SL          17.543627 15.80403 17.62931 15.718086 16.842947 17.56647
#> Lincoln Continental  8.919666 10.46411 11.49057  9.058440  7.248073 10.79303
#> Chrysler Imperial    9.022469 10.89780 11.52304  9.027576  7.329615 10.79875
#> AMC Javelin         18.631889 18.62255 21.02658 19.536933 18.273147 19.49886
#> Fiat X1-9           27.574195 33.08942 29.12801 28.153534 28.356123 26.17803
#> Ford Pantera L      16.547624 21.15062 17.82215 17.260680 17.111490 18.02345
#>                         boot7    boot8     boot9   boot10    boot11   boot12
#> Mazda RX4           22.486255 21.38286 21.610324 23.08422 24.291342 20.93438
#> Valiant             19.154842 20.97607 20.754243 20.23630 18.027761 19.43478
#> Duster 360          17.034553 12.51162 16.187325 14.41893 15.236248 14.30810
#> Merc 230            21.204176 24.06061 24.789197 22.64368 22.967191 24.07882
#> Merc 450SL          16.999073 15.48568 16.978615 16.86316 17.385591 16.33228
#> Lincoln Continental  8.374911 12.45827  8.520507 11.46582  9.798639 10.79223
#> Chrysler Imperial    8.622100 12.62926  8.558614 11.27337  9.592298 10.89383
#> AMC Javelin         18.547813 17.50024 18.212283 18.43781 18.627835 20.53950
#> Fiat X1-9           26.640025 32.01573 27.342741 26.48623 31.368523 25.85582
#> Ford Pantera L      18.681975 21.25797 17.006852 15.14260 18.560905 17.44830
#>                       boot13    boot14   boot15   boot16    boot17    boot18
#> Mazda RX4           22.03864 22.651059 20.82898 22.37334 23.212715 22.232274
#> Valiant             18.67078 18.984169 18.39854 22.05235 19.130710 19.519889
#> Duster 360          16.93199 16.366149 15.47038 11.08385 15.565383 14.929037
#> Merc 230            24.09137 20.680635 23.55040 28.93754 20.588928 22.349134
#> Merc 450SL          17.09156 16.748638 15.94579 15.80868 16.476548 17.826066
#> Lincoln Continental 11.10656  7.700033 10.05250 13.82411  9.709705  7.886874
#> Chrysler Imperial   11.32578  7.901002 10.31644 13.21525  9.647786  7.850047
#> AMC Javelin         18.45623 18.617316 18.90448 17.98194 18.197511 19.800347
#> Fiat X1-9           28.49587 26.893450 25.11489 30.44060 26.764735 28.159028
#> Ford Pantera L      18.62032 18.082208 16.49400 15.20114 17.778348 16.140974
#>                        boot19   boot20
#> Mazda RX4           23.192663 25.30041
#> Valiant             20.554695 20.10232
#> Duster 360          16.313202 15.28669
#> Merc 230            27.046948 21.32139
#> Merc 450SL          19.301097 17.03815
#> Lincoln Continental  9.059540 10.70861
#> Chrysler Imperial    8.805335 10.48920
#> AMC Javelin         20.848505 18.81698
#> Fiat X1-9           29.033780 28.91711
#> Ford Pantera L      16.812706 19.02166

## Alternal Matrix Interface

# Train model
set.seed(123)
bolasso_mat <- bolasso(
  x = model.matrix(mpg ~ . - 1, mtcars_train),
  y = mtcars_train[, 1],
  data = mtcars_train,
  n.boot = 20,
  nfolds = 5,
  implement = "glmnet"
)

# Extract selected variables
selected_vars(bolasso_mat, threshold = 0.9, select = "lambda.min")
#> # A tibble: 2 × 2
#>   variable  mean_coef
#>   <chr>         <dbl>
#> 1 Intercept     23.1 
#> 2 wt            -3.68

# Bagged ensemble prediction on test data
predict(bolasso_mat,
        new.data = model.matrix(mpg ~ . - 1, mtcars_test),
        select = "lambda.min")
#>                         boot1    boot2    boot3     boot4     boot5    boot6
#> Mazda RX4           21.202865 24.06137 20.37871 19.693430 22.115000 22.88248
#> Valiant             20.825762 20.27784 14.05869 21.108939 20.806772 19.94063
#> Duster 360          15.241608 12.73562 11.37251 16.573997 16.048799 16.93248
#> Merc 230            27.293281 23.38017 26.61376 27.069320 25.023952 21.55425
#> Merc 450SL          17.543627 15.80403 17.62931 15.718086 16.842947 17.56647
#> Lincoln Continental  8.919666 10.46411 11.49057  9.058440  7.248073 10.79303
#> Chrysler Imperial    9.022469 10.89780 11.52304  9.027576  7.329615 10.79875
#> AMC Javelin         18.631889 18.62255 21.02658 19.536933 18.273147 19.49886
#> Fiat X1-9           27.574195 33.08942 29.12801 28.153534 28.356123 26.17803
#> Ford Pantera L      16.547624 21.15062 17.82215 17.260680 17.111490 18.02345
#>                         boot7    boot8     boot9   boot10    boot11   boot12
#> Mazda RX4           22.486255 21.38286 21.610324 23.08422 24.291342 20.93438
#> Valiant             19.154842 20.97607 20.754243 20.23630 18.027761 19.43478
#> Duster 360          17.034553 12.51162 16.187325 14.41893 15.236248 14.30810
#> Merc 230            21.204176 24.06061 24.789197 22.64368 22.967191 24.07882
#> Merc 450SL          16.999073 15.48568 16.978615 16.86316 17.385591 16.33228
#> Lincoln Continental  8.374911 12.45827  8.520507 11.46582  9.798639 10.79223
#> Chrysler Imperial    8.622100 12.62926  8.558614 11.27337  9.592298 10.89383
#> AMC Javelin         18.547813 17.50024 18.212283 18.43781 18.627835 20.53950
#> Fiat X1-9           26.640025 32.01573 27.342741 26.48623 31.368523 25.85582
#> Ford Pantera L      18.681975 21.25797 17.006852 15.14260 18.560905 17.44830
#>                       boot13    boot14   boot15   boot16    boot17    boot18
#> Mazda RX4           22.03864 22.651059 20.82898 22.37334 23.212715 22.232274
#> Valiant             18.67078 18.984169 18.39854 22.05235 19.130710 19.519889
#> Duster 360          16.93199 16.366149 15.47038 11.08385 15.565383 14.929037
#> Merc 230            24.09137 20.680635 23.55040 28.93754 20.588928 22.349134
#> Merc 450SL          17.09156 16.748638 15.94579 15.80868 16.476548 17.826066
#> Lincoln Continental 11.10656  7.700033 10.05250 13.82411  9.709705  7.886874
#> Chrysler Imperial   11.32578  7.901002 10.31644 13.21525  9.647786  7.850047
#> AMC Javelin         18.45623 18.617316 18.90448 17.98194 18.197511 19.800347
#> Fiat X1-9           28.49587 26.893450 25.11489 30.44060 26.764735 28.159028
#> Ford Pantera L      18.62032 18.082208 16.49400 15.20114 17.778348 16.140974
#>                        boot19   boot20
#> Mazda RX4           23.192663 25.30041
#> Valiant             20.554695 20.10232
#> Duster 360          16.313202 15.28669
#> Merc 230            27.046948 21.32139
#> Merc 450SL          19.301097 17.03815
#> Lincoln Continental  9.059540 10.70861
#> Chrysler Imperial    8.805335 10.48920
#> AMC Javelin         20.848505 18.81698
#> Fiat X1-9           29.033780 28.91711
#> Ford Pantera L      16.812706 19.02166