Bootsrap-enhanced Lasso

This function implements model-consistent Lasso estimation through the bootstrap. It supports parallel processing by way of the future package, allowing the user to flexibly specify many parallelization methods. This method was developed as a variable-selection algorithm, but this package also supports making ensemble predictions on new data using the bagged Lasso models.

Usage

bolasso(
  form,
  data,
  n.boot = 100,
  progress = TRUE,
  implement = "glmnet",
  x = NULL,
  y = NULL,
  ...
)

Arguments

form

An optional object of class formula (or one that can be coerced to that class): a symbolic description of the model to be fitted. Can be omitted when x and y are non-missing.

data

An optional object of class data.frame that contains the modeling variables referenced in form. Can be omitted when x and y are non-missing.

n.boot

An integer specifying the number of bootstrap replicates.

progress

A boolean indicating whether to display progress across bootstrap folds.

implement

A character; either 'glmnet' or 'gamlr', specifying which Lasso implementation to utilize. For specific modeling details, see glmnet::cv.glmnet or gamlr::cv.gamlr.

x

An optional predictor matrix in lieu of form and data.

y

An optional response vector in lieu of form and data.

...

Additional parameters to pass to either glmnet::cv.glmnet or gamlr::cv.gamlr.

Value

An object of class bolasso. This object is a list of length n.boot of cv.glmnet or cv.gamlr objects.

References

Bach FR (2008). “Bolasso: model consistent Lasso estimation through the bootstrap.” CoRR, abs/0804.1302. 0804.1302, http://arxiv.org/abs/0804.1302.

See also

Examples

mtcars[, c(2, 10:11)] <- lapply(mtcars[, c(2, 10:11)], as.factor)
idx <- sample(nrow(mtcars), 22)
mtcars_train <- mtcars[idx, ]
mtcars_test <- mtcars[-idx, ]

## Formula Interface

# Train model
set.seed(123)
bolasso_form <- bolasso(
  form = mpg ~ .,
  data = mtcars_train,
  n.boot = 20,
  nfolds = 5,
  implement = "glmnet"
)
#> Loaded glmnet 4.1-3

# Extract selected variables
selected_vars(bolasso_form, threshold = 0.9, select = "lambda.min")
#> # A tibble: 3 × 2
#>   variable  mean_coef
#>   <chr>         <dbl>
#> 1 Intercept   27.2   
#> 2 hp          -0.0471
#> 3 wt          -4.08  

# Bagged ensemble prediction on test data
predict(bolasso_form,
        new.data = mtcars_test,
        select = "lambda.min")
#>                         boot1     boot2     boot3     boot4    boot5     boot6
#> Mazda RX4 Wag       23.644560 21.388571 20.358116 20.972528 22.68057 20.787645
#> Lincoln Continental  8.784877  8.597955  9.343796  8.777839 11.19919  6.840606
#> Chrysler Imperial    8.171548  8.289624  9.388172  8.656452 10.93344  6.836580
#> Toyota Corona       26.752496 31.351424 25.791595 27.184781 25.96792 27.023569
#> Dodge Challenger    19.120912 18.659944 18.060018 17.468552 19.31797 17.603522
#> AMC Javelin         19.513047 22.985234 18.370341 18.487985 19.58160 18.363549
#> Camaro Z28          14.136871 12.674783 14.638664 13.842982 15.09053 13.517010
#> Pontiac Firebird    15.991929 16.893282 16.466754 16.008954 17.45866 15.683764
#> Fiat X1-9           33.007133 31.701820 28.230900 28.473714 28.66737 29.488317
#> Maserati Bora       11.034076 -6.817640 14.867625 14.172756 12.86328 13.788139
#>                         boot7     boot8    boot9    boot10    boot11    boot12
#> Mazda RX4 Wag       20.586708 20.560185 21.01413 22.163923 19.666553 20.829037
#> Lincoln Continental  9.529867  9.411719 12.13900  7.683959  8.713384  7.994905
#> Chrysler Imperial    9.354100  9.102798 11.85880  7.593633  7.997630  7.312644
#> Toyota Corona       27.050695 25.089718 28.94630 25.908285 30.415963 31.995634
#> Dodge Challenger    17.596372 18.991315 18.66873 18.676292 15.702424 17.329441
#> AMC Javelin         18.551067 19.285890 19.01541 19.101450 15.690219 18.633026
#> Camaro Z28          13.862294 13.735809 17.39459 13.931444 14.168614 11.713639
#> Pontiac Firebird    15.913358 16.893828 16.33920 16.070140 12.277919 15.194288
#> Fiat X1-9           29.074728 28.759828 32.64050 29.426901 33.322289 30.868247
#> Maserati Bora       11.125636 11.804403 11.50749 12.209294  8.360078  9.479482
#>                       boot13    boot14    boot15    boot16    boot17   boot18
#> Mazda RX4 Wag       20.81270 20.732359 20.440081 20.156246 15.886862 20.87350
#> Lincoln Continental 10.52967  9.941445  9.856206  4.975690  8.029345 12.14497
#> Chrysler Imperial   10.41874 10.024477 10.966962  4.961956  6.781118 12.12705
#> Toyota Corona       24.14251 28.063701 24.927970 26.518899 28.989700 28.52693
#> Dodge Challenger    17.89458 17.493347 16.080131 17.534670 15.077864 17.76562
#> AMC Javelin         18.15507 18.209491 21.038380 18.387893 17.010905 18.01619
#> Camaro Z28          14.67799 14.978282 19.420946 12.735231  9.498416 15.05392
#> Pontiac Firebird    16.31017 16.047028 14.428595 15.262942 12.711066 16.08315
#> Fiat X1-9           26.49638 29.532660 32.095662 28.438037 30.104215 30.36523
#> Maserati Bora       13.38722 15.625099 -5.575738 12.886280  7.430002 14.05939
#>                        boot19    boot20
#> Mazda RX4 Wag       20.499438 22.593902
#> Lincoln Continental  8.424775 10.029108
#> Chrysler Imperial    8.240081  9.836698
#> Toyota Corona       28.630112 26.232080
#> Dodge Challenger    17.501705 18.986417
#> AMC Javelin         18.206477 19.293807
#> Camaro Z28          14.506161 14.801208
#> Pontiac Firebird    15.329656 17.014267
#> Fiat X1-9           31.445301 29.136823
#> Maserati Bora       13.979657 12.909015

## Alternal Matrix Interface

# Train model
set.seed(123)
bolasso_mat <- bolasso(
  x = model.matrix(mpg ~ . - 1, mtcars_train),
  y = mtcars_train[, 1],
  data = mtcars_train,
  n.boot = 20,
  nfolds = 5,
  implement = "glmnet"
)

# Extract selected variables
selected_vars(bolasso_mat, threshold = 0.9, select = "lambda.min")
#> # A tibble: 3 × 2
#>   variable  mean_coef
#>   <chr>         <dbl>
#> 1 Intercept   27.2   
#> 2 hp          -0.0471
#> 3 wt          -4.08  

# Bagged ensemble prediction on test data
predict(bolasso_mat,
        new.data = model.matrix(mpg ~ . - 1, mtcars_test),
        select = "lambda.min")
#>                         boot1     boot2     boot3     boot4    boot5     boot6
#> Mazda RX4 Wag       23.644560 21.388571 20.358116 20.972528 22.68057 20.787645
#> Lincoln Continental  8.784877  8.597955  9.343796  8.777839 11.19919  6.840606
#> Chrysler Imperial    8.171548  8.289624  9.388172  8.656452 10.93344  6.836580
#> Toyota Corona       26.752496 31.351424 25.791595 27.184781 25.96792 27.023569
#> Dodge Challenger    19.120912 18.659944 18.060018 17.468552 19.31797 17.603522
#> AMC Javelin         19.513047 22.985234 18.370341 18.487985 19.58160 18.363549
#> Camaro Z28          14.136871 12.674783 14.638664 13.842982 15.09053 13.517010
#> Pontiac Firebird    15.991929 16.893282 16.466754 16.008954 17.45866 15.683764
#> Fiat X1-9           33.007133 31.701820 28.230900 28.473714 28.66737 29.488317
#> Maserati Bora       11.034076 -6.817640 14.867625 14.172756 12.86328 13.788139
#>                         boot7     boot8    boot9    boot10    boot11    boot12
#> Mazda RX4 Wag       20.586708 20.560185 21.01413 22.163923 19.666553 20.829037
#> Lincoln Continental  9.529867  9.411719 12.13900  7.683959  8.713384  7.994905
#> Chrysler Imperial    9.354100  9.102798 11.85880  7.593633  7.997630  7.312644
#> Toyota Corona       27.050695 25.089718 28.94630 25.908285 30.415963 31.995634
#> Dodge Challenger    17.596372 18.991315 18.66873 18.676292 15.702424 17.329441
#> AMC Javelin         18.551067 19.285890 19.01541 19.101450 15.690219 18.633026
#> Camaro Z28          13.862294 13.735809 17.39459 13.931444 14.168614 11.713639
#> Pontiac Firebird    15.913358 16.893828 16.33920 16.070140 12.277919 15.194288
#> Fiat X1-9           29.074728 28.759828 32.64050 29.426901 33.322289 30.868247
#> Maserati Bora       11.125636 11.804403 11.50749 12.209294  8.360078  9.479482
#>                       boot13    boot14    boot15    boot16    boot17   boot18
#> Mazda RX4 Wag       20.81270 20.732359 20.440081 20.156246 15.886862 20.87350
#> Lincoln Continental 10.52967  9.941445  9.856206  4.975690  8.029345 12.14497
#> Chrysler Imperial   10.41874 10.024477 10.966962  4.961956  6.781118 12.12705
#> Toyota Corona       24.14251 28.063701 24.927970 26.518899 28.989700 28.52693
#> Dodge Challenger    17.89458 17.493347 16.080131 17.534670 15.077864 17.76562
#> AMC Javelin         18.15507 18.209491 21.038380 18.387893 17.010905 18.01619
#> Camaro Z28          14.67799 14.978282 19.420946 12.735231  9.498416 15.05392
#> Pontiac Firebird    16.31017 16.047028 14.428595 15.262942 12.711066 16.08315
#> Fiat X1-9           26.49638 29.532660 32.095662 28.438037 30.104215 30.36523
#> Maserati Bora       13.38722 15.625099 -5.575738 12.886280  7.430002 14.05939
#>                        boot19    boot20
#> Mazda RX4 Wag       20.499438 22.593902
#> Lincoln Continental  8.424775 10.029108
#> Chrysler Imperial    8.240081  9.836698
#> Toyota Corona       28.630112 26.232080
#> Dodge Challenger    17.501705 18.986417
#> AMC Javelin         18.206477 19.293807
#> Camaro Z28          14.506161 14.801208
#> Pontiac Firebird    15.329656 17.014267
#> Fiat X1-9           31.445301 29.136823
#> Maserati Bora       13.979657 12.909015