Bootsrap-enhanced Lasso

This function implements model-consistent Lasso estimation through the bootstrap. It supports parallel processing by way of the future package, allowing the user to flexibly specify many parallelization methods. This method was developed as a variable-selection algorithm, but this package also supports making ensemble predictions on new data using the bagged Lasso models.

Usage

bolasso(
  form,
  data,
  n.boot = 100,
  progress = TRUE,
  implement = "glmnet",
  x = NULL,
  y = NULL,
  ...
)

Arguments

form

An optional object of class formula (or one that can be coerced to that class): a symbolic description of the model to be fitted. Can be omitted when x and y are non-missing.

data

An optional object of class data.frame that contains the modeling variables referenced in form. Can be omitted when x and y are non-missing.

n.boot

An integer specifying the number of bootstrap replicates.

progress

A boolean indicating whether to display progress across bootstrap folds.

implement

A character; either 'glmnet' or 'gamlr', specifying which Lasso implementation to utilize. For specific modeling details, see glmnet::cv.glmnet or gamlr::cv.gamlr.

x

An optional predictor matrix in lieu of form and data.

y

An optional response vector in lieu of form and data.

...

Additional parameters to pass to either glmnet::cv.glmnet or gamlr::cv.gamlr.

Value

An object of class bolasso. This object is a list of length n.boot of cv.glmnet or cv.gamlr objects.

References

Bach FR (2008). “Bolasso: model consistent Lasso estimation through the bootstrap.” CoRR, abs/0804.1302. 0804.1302, https://arxiv.org/abs/0804.1302.

See also

Examples

mtcars[, c(2, 10:11)] <- lapply(mtcars[, c(2, 10:11)], as.factor)
idx <- sample(nrow(mtcars), 22)
mtcars_train <- mtcars[idx, ]
mtcars_test <- mtcars[-idx, ]

## Formula Interface

# Train model
set.seed(123)
bolasso_form <- bolasso(
  form = mpg ~ .,
  data = mtcars_train,
  n.boot = 20,
  nfolds = 5,
  implement = "glmnet"
)
#> Loaded glmnet 4.1-3

# Extract selected variables
selected_vars(bolasso_form, threshold = 0.9, select = "lambda.min")
#> # A tibble: 1 × 2
#>   variable  mean_coef
#>   <chr>         <dbl>
#> 1 Intercept      20.6

# Bagged ensemble prediction on test data
predict(bolasso_form,
        new.data = mtcars_test,
        select = "lambda.min")
#>                     boot1    boot2    boot3    boot4    boot5    boot6    boot7
#> Datsun 710       27.31744 35.30460 22.56104 24.94733 28.11741 24.65981 25.12225
#> Hornet 4 Drive   17.68265 21.99839 18.41801 19.68089 17.78921 19.03875 19.16741
#> Valiant          17.41601 21.54064 18.48137 18.78526 17.53842 17.88850 18.22299
#> Fiat 128         29.45379 33.60626 23.13963 26.11077 29.53927 25.81841 25.89131
#> Dodge Challenger 15.29591 16.21312 16.48303 16.87129 15.33608 17.10574 17.51063
#> Pontiac Firebird 14.09309 18.18344 15.14067 15.07250 14.02199 16.78481 15.92861
#> Fiat X1-9        30.01751 33.54975 23.45426 27.22857 30.08845 26.36628 26.96912
#> Porsche 914-2    19.85540 21.60146 23.17706 25.36662 21.95092 26.36064 24.92076
#> Ford Pantera L   18.01434 21.60653 16.93380 16.83211 16.39334 19.43155 17.74771
#> Ferrari Dino     15.35453 29.42953 20.22788 20.04708 16.52064 20.21708 20.30083
#>                     boot8    boot9   boot10   boot11    boot12   boot13
#> Datsun 710       30.48082 25.88657 29.13457 27.53030 32.775602 23.24962
#> Hornet 4 Drive   21.64802 18.88601 18.13170 20.46104 15.958540 19.48611
#> Valiant          21.64802 17.89146 17.52773 20.19147 15.113695 18.64267
#> Fiat 128         30.48082 26.82911 30.06774 28.51041 33.414104 24.86015
#> Dodge Challenger 16.81288 18.23924 16.49061 18.67685  9.764226 17.65523
#> Pontiac Firebird 16.81288 16.43075 15.76034 17.10994  9.517604 16.81455
#> Fiat X1-9        30.48082 27.98849 30.47069 29.06260 33.676636 25.30901
#> Porsche 914-2    25.22793 25.79840 30.19871 27.90191 29.466284 30.58344
#> Ford Pantera L   20.97459 18.61680 17.89895 16.65995 20.167448 19.96676
#> Ferrari Dino     23.86654 20.43695 19.63554 21.35871 22.607166 24.22750
#>                    boot14   boot15    boot16    boot17   boot18   boot19
#> Datsun 710       28.97178 26.08649 26.556876 33.043373 28.54322 24.09434
#> Hornet 4 Drive   18.30869 21.20926  7.213108 18.640396 19.98645 17.36297
#> Valiant          17.72828 20.66452  1.746347 17.795677 19.37794 16.68577
#> Fiat 128         29.90905 27.30683 28.227967 32.942777 29.36361 24.74215
#> Dodge Challenger 14.36743 19.02695 10.553663 16.594709 16.03645 16.14649
#> Pontiac Firebird 13.40772 17.28010 15.154103 17.617337 14.66817 15.52967
#> Fiat X1-9        30.43328 28.07116 28.833386 33.067162 30.11772 25.24896
#> Porsche 914-2    27.41566 26.66295 40.760130  8.537459 26.71836 23.13089
#> Ford Pantera L   17.60965 17.26113 32.236810  5.802800 17.75464 17.48860
#> Ferrari Dino     19.86385 21.29655 22.586980 10.310645 20.47154 18.57458
#>                    boot20
#> Datsun 710       32.48084
#> Hornet 4 Drive   12.52993
#> Valiant          12.00595
#> Fiat 128         32.85745
#> Dodge Challenger 15.62167
#> Pontiac Firebird 16.14566
#> Fiat X1-9        32.85745
#> Porsche 914-2    22.25654
#> Ford Pantera L   13.32996
#> Ferrari Dino     12.06990

## Alternal Matrix Interface

# Train model
set.seed(123)
bolasso_mat <- bolasso(
  x = model.matrix(mpg ~ . - 1, mtcars_train),
  y = mtcars_train[, 1],
  data = mtcars_train,
  n.boot = 20,
  nfolds = 5,
  implement = "glmnet"
)

# Extract selected variables
selected_vars(bolasso_mat, threshold = 0.9, select = "lambda.min")
#> # A tibble: 1 × 2
#>   variable  mean_coef
#>   <chr>         <dbl>
#> 1 Intercept      20.6

# Bagged ensemble prediction on test data
predict(bolasso_mat,
        new.data = model.matrix(mpg ~ . - 1, mtcars_test),
        select = "lambda.min")
#>                     boot1    boot2    boot3    boot4    boot5    boot6    boot7
#> Datsun 710       27.31744 35.30460 22.56104 24.94733 28.11741 24.65981 25.12225
#> Hornet 4 Drive   17.68265 21.99839 18.41801 19.68089 17.78921 19.03875 19.16741
#> Valiant          17.41601 21.54064 18.48137 18.78526 17.53842 17.88850 18.22299
#> Fiat 128         29.45379 33.60626 23.13963 26.11077 29.53927 25.81841 25.89131
#> Dodge Challenger 15.29591 16.21312 16.48303 16.87129 15.33608 17.10574 17.51063
#> Pontiac Firebird 14.09309 18.18344 15.14067 15.07250 14.02199 16.78481 15.92861
#> Fiat X1-9        30.01751 33.54975 23.45426 27.22857 30.08845 26.36628 26.96912
#> Porsche 914-2    19.85540 21.60146 23.17706 25.36662 21.95092 26.36064 24.92076
#> Ford Pantera L   18.01434 21.60653 16.93380 16.83211 16.39334 19.43155 17.74771
#> Ferrari Dino     15.35453 29.42953 20.22788 20.04708 16.52064 20.21708 20.30083
#>                     boot8    boot9   boot10   boot11    boot12   boot13
#> Datsun 710       30.48082 25.88657 29.13457 27.53030 32.775602 23.24962
#> Hornet 4 Drive   21.64802 18.88601 18.13170 20.46104 15.958540 19.48611
#> Valiant          21.64802 17.89146 17.52773 20.19147 15.113695 18.64267
#> Fiat 128         30.48082 26.82911 30.06774 28.51041 33.414104 24.86015
#> Dodge Challenger 16.81288 18.23924 16.49061 18.67685  9.764226 17.65523
#> Pontiac Firebird 16.81288 16.43075 15.76034 17.10994  9.517604 16.81455
#> Fiat X1-9        30.48082 27.98849 30.47069 29.06260 33.676636 25.30901
#> Porsche 914-2    25.22793 25.79840 30.19871 27.90191 29.466284 30.58344
#> Ford Pantera L   20.97459 18.61680 17.89895 16.65995 20.167448 19.96676
#> Ferrari Dino     23.86654 20.43695 19.63554 21.35871 22.607166 24.22750
#>                    boot14   boot15    boot16    boot17   boot18   boot19
#> Datsun 710       28.97178 26.08649 26.556876 33.043373 28.54322 24.09434
#> Hornet 4 Drive   18.30869 21.20926  7.213108 18.640396 19.98645 17.36297
#> Valiant          17.72828 20.66452  1.746347 17.795677 19.37794 16.68577
#> Fiat 128         29.90905 27.30683 28.227967 32.942777 29.36361 24.74215
#> Dodge Challenger 14.36743 19.02695 10.553663 16.594709 16.03645 16.14649
#> Pontiac Firebird 13.40772 17.28010 15.154103 17.617337 14.66817 15.52967
#> Fiat X1-9        30.43328 28.07116 28.833386 33.067162 30.11772 25.24896
#> Porsche 914-2    27.41566 26.66295 40.760130  8.537459 26.71836 23.13089
#> Ford Pantera L   17.60965 17.26113 32.236810  5.802800 17.75464 17.48860
#> Ferrari Dino     19.86385 21.29655 22.586980 10.310645 20.47154 18.57458
#>                    boot20
#> Datsun 710       32.48084
#> Hornet 4 Drive   12.52993
#> Valiant          12.00595
#> Fiat 128         32.85745
#> Dodge Challenger 15.62167
#> Pontiac Firebird 16.14566
#> Fiat X1-9        32.85745
#> Porsche 914-2    22.25654
#> Ford Pantera L   13.32996
#> Ferrari Dino     12.06990