Bootsrap-enhanced Lasso

This function implements model-consistent Lasso estimation through the bootstrap. It supports parallel processing by way of the future package, allowing the user to flexibly specify many parallelization methods. This method was developed as a variable-selection algorithm, but this package also supports making ensemble predictions on new data using the bagged Lasso models.

Usage

bolasso(
  form,
  data,
  n.boot = 100,
  progress = TRUE,
  implement = "glmnet",
  x = NULL,
  y = NULL,
  ...
)

Arguments

form

An optional object of class formula (or one that can be coerced to that class): a symbolic description of the model to be fitted. Can be omitted when x and y are non-missing.

data

An optional object of class data.frame that contains the modeling variables referenced in form. Can be omitted when x and y are non-missing.

n.boot

An integer specifying the number of bootstrap replicates.

progress

A boolean indicating whether to display progress across bootstrap folds.

implement

A character; either 'glmnet' or 'gamlr', specifying which Lasso implementation to utilize. For specific modeling details, see glmnet::cv.glmnet or gamlr::cv.gamlr.

x

An optional predictor matrix in lieu of form and data.

y

An optional response vector in lieu of form and data.

...

Additional parameters to pass to either glmnet::cv.glmnet or gamlr::cv.gamlr.

Value

An object of class bolasso. This object is a list of length n.boot of cv.glmnet or cv.gamlr objects.

References

Bach FR (2008). “Bolasso: model consistent Lasso estimation through the bootstrap.” CoRR, abs/0804.1302. 0804.1302, https://arxiv.org/abs/0804.1302.

See also

Examples

mtcars[, c(2, 10:11)] <- lapply(mtcars[, c(2, 10:11)], as.factor)
idx <- sample(nrow(mtcars), 22)
mtcars_train <- mtcars[idx, ]
mtcars_test <- mtcars[-idx, ]

## Formula Interface

# Train model
set.seed(123)
bolasso_form <- bolasso(
  form = mpg ~ .,
  data = mtcars_train,
  n.boot = 20,
  nfolds = 5,
  implement = "glmnet"
)
#> Loaded glmnet 4.1-3

# Extract selected variables
selected_vars(bolasso_form, threshold = 0.9, select = "lambda.min")
#> # A tibble: 2 × 2
#>   variable  mean_coef
#>   <chr>         <dbl>
#> 1 Intercept     23.9 
#> 2 wt            -2.80

# Bagged ensemble prediction on test data
predict(bolasso_form,
        new.data = mtcars_test,
        select = "lambda.min")
#>                     boot1    boot2    boot3    boot4    boot5    boot6    boot7
#> Mazda RX4        24.32901 24.21034 20.91563 21.20545 21.53248 24.83083 22.50248
#> Valiant          20.65363 21.82481 19.43796 18.06856 23.38860 18.63808 16.71835
#> Duster 360       12.82107 15.42490 13.39675 13.67175 13.96376 16.11449 14.71800
#> Merc 240D        23.85432 27.32492 21.09902 21.79485 25.81947 22.52448 22.01140
#> Merc 280         22.03263 21.80920 20.83275 18.64068 22.92831 20.40705 20.67178
#> Merc 280C        22.09024 21.89185 20.83275 19.30490 24.06370 20.40705 20.67178
#> Pontiac Firebird 14.21702 16.97009 16.02204 14.95021 14.33608 15.71402 15.44842
#> Fiat X1-9        29.00007 31.22524 22.93388 26.59674 29.19818 28.90026 25.79914
#> Ford Pantera L   14.67974 15.75559 18.81189 19.46314 19.88839 17.97249 23.65535
#> Maserati Bora     8.17903 11.92126 19.33627 16.83822 17.70453 14.95881 18.60074
#>                     boot8    boot9   boot10   boot11   boot12   boot13   boot14
#> Mazda RX4        24.53382 22.64985 21.60617 24.10523 25.48787 23.13283 24.39220
#> Valiant          21.81221 19.63185 20.46858 18.72015 21.68861 19.84277 19.93881
#> Duster 360       14.43522 16.04412 14.26196 16.31534 14.12367 16.82259 14.63875
#> Merc 240D        24.10495 18.60361 25.50321 27.03165 25.70761 21.70702 23.46948
#> Merc 280         21.19416 15.92138 20.82638 26.22539 22.46531 19.65698 22.76064
#> Merc 280C        21.19416 15.92138 20.82638 26.22539 22.46531 19.65698 22.76064
#> Pontiac Firebird 16.33217 13.42070 16.73453 16.17236 18.05692 16.66355 14.29783
#> Fiat X1-9        29.81751 33.08417 27.44534 30.61290 28.37560 27.25552 28.54190
#> Ford Pantera L   17.39372 14.22523 16.19651 21.18934 16.88868 18.32005 19.26443
#> Maserati Bora    13.72572 14.73164 12.97259 18.53552 11.61064 15.44666 18.56064
#>                    boot15   boot16   boot17   boot18   boot19   boot20
#> Mazda RX4        22.42991 21.57306 25.60223 24.71839 23.36276 22.42462
#> Valiant          12.79586 18.51480 19.82291 19.59867 18.99358 19.90148
#> Duster 360       19.77079 18.04841 22.92488 12.87120 13.55626 13.73338
#> Merc 240D        21.67654 22.23869 19.82389 23.87678 22.74378 23.86378
#> Merc 280         16.69549 18.57229 21.23633 21.31540 19.98058 21.50879
#> Merc 280C        16.69549 18.57229 21.23633 21.31540 19.98058 21.50879
#> Pontiac Firebird 15.37696 19.91437 15.08482 15.51816 15.67361 15.76164
#> Fiat X1-9        32.40950 23.95723 33.31007 28.04523 26.79976 25.51103
#> Ford Pantera L   26.06145 19.58420 26.13294 16.72805 15.84442 14.77116
#> Maserati Bora    24.54967 18.16352 22.92323 12.64562 11.38454 11.28743

## Alternal Matrix Interface

# Train model
set.seed(123)
bolasso_mat <- bolasso(
  x = model.matrix(mpg ~ . - 1, mtcars_train),
  y = mtcars_train[, 1],
  data = mtcars_train,
  n.boot = 20,
  nfolds = 5,
  implement = "glmnet"
)

# Extract selected variables
selected_vars(bolasso_mat, threshold = 0.9, select = "lambda.min")
#> # A tibble: 2 × 2
#>   variable  mean_coef
#>   <chr>         <dbl>
#> 1 Intercept     23.9 
#> 2 wt            -2.80

# Bagged ensemble prediction on test data
predict(bolasso_mat,
        new.data = model.matrix(mpg ~ . - 1, mtcars_test),
        select = "lambda.min")
#>                     boot1    boot2    boot3    boot4    boot5    boot6    boot7
#> Mazda RX4        24.32901 24.21034 20.91563 21.20545 21.53128 24.83083 22.50248
#> Valiant          20.65363 21.82481 19.43796 18.06856 23.38921 18.63808 16.71835
#> Duster 360       12.82107 15.42490 13.39675 13.67175 13.96445 16.11449 14.71800
#> Merc 240D        23.85432 27.32492 21.09902 21.79485 25.81913 22.52448 22.01140
#> Merc 280         22.03263 21.80920 20.83275 18.64068 22.92760 20.40705 20.67178
#> Merc 280C        22.09024 21.89185 20.83275 19.30490 24.06339 20.40705 20.67178
#> Pontiac Firebird 14.21702 16.97009 16.02204 14.95021 14.33643 15.71402 15.44842
#> Fiat X1-9        29.00007 31.22524 22.93388 26.59674 29.19818 28.90026 25.79914
#> Ford Pantera L   14.67974 15.75559 18.81189 19.46314 19.88884 17.97249 23.65535
#> Maserati Bora     8.17903 11.92126 19.33627 16.83822 17.70539 14.95881 18.60074
#>                     boot8    boot9   boot10   boot11   boot12   boot13   boot14
#> Mazda RX4        24.53382 22.64985 21.60617 24.10523 25.48787 23.13283 24.39220
#> Valiant          21.81221 19.63185 20.46858 18.72015 21.68861 19.84277 19.93881
#> Duster 360       14.43522 16.04412 14.26196 16.31534 14.12367 16.82259 14.63875
#> Merc 240D        24.10495 18.60361 25.50321 27.03165 25.70761 21.70702 23.46948
#> Merc 280         21.19416 15.92138 20.82638 26.22539 22.46531 19.65698 22.76064
#> Merc 280C        21.19416 15.92138 20.82638 26.22539 22.46531 19.65698 22.76064
#> Pontiac Firebird 16.33217 13.42070 16.73453 16.17236 18.05692 16.66355 14.29783
#> Fiat X1-9        29.81751 33.08417 27.44534 30.61290 28.37560 27.25552 28.54190
#> Ford Pantera L   17.39372 14.22523 16.19651 21.18934 16.88868 18.32005 19.26443
#> Maserati Bora    13.72572 14.73164 12.97259 18.53552 11.61064 15.44666 18.56064
#>                    boot15   boot16   boot17   boot18   boot19   boot20
#> Mazda RX4        22.42991 21.57306 25.60223 24.71839 23.36276 22.42462
#> Valiant          12.79586 18.51480 19.82291 19.59867 18.99358 19.90148
#> Duster 360       19.77079 18.04841 22.92488 12.87120 13.55626 13.73338
#> Merc 240D        21.67654 22.23869 19.82389 23.87678 22.74378 23.86378
#> Merc 280         16.69549 18.57229 21.23633 21.31540 19.98058 21.50879
#> Merc 280C        16.69549 18.57229 21.23633 21.31540 19.98058 21.50879
#> Pontiac Firebird 15.37696 19.91437 15.08482 15.51816 15.67361 15.76164
#> Fiat X1-9        32.40950 23.95723 33.31007 28.04523 26.79976 25.51103
#> Ford Pantera L   26.06145 19.58420 26.13294 16.72805 15.84442 14.77116
#> Maserati Bora    24.54967 18.16352 22.92323 12.64562 11.38454 11.28743